% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coverage.R
\name{calc_coverage}
\alias{calc_coverage}
\title{calc_coverage}
\usage{
calc_coverage(
  get_isdb = NULL,
  get_marfis = NULL,
  marfMatchOnly = TRUE,
  agg.poly.shp = NULL,
  agg.poly.field = NULL,
  ...
)
}
\arguments{
\item{get_isdb}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_isdb()} function - it contains dataframes of both the
trip and set information from ISDB related to the specified fleet}

\item{get_marfis}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_marfis()} function - it contains dataframes of both the
trip and set information from MARFIS related to the specified fleet}

\item{marfMatchOnly}{default is \code{TRUE}.  This determines whether the
function should be run on all ISDB data, or just on those that have been
matched to MARFIS data (the default).}

\item{agg.poly.shp}{default is \code{NULL}.  This is the shapefile that has
polygons that should be checked for sufficient unique values of the
sens.fields.  If NULL, NAFO zones will be used.  Otherwise, a path to any
polygon shapefile can be provided.}

\item{agg.poly.field}{default is \code{NULL}.  This identifies the field within
the shapefile provided to agg.poly.shp that should be used to check for
sufficient unique values of the sens.fields.}

\item{...}{other arguments passed to methods}
}
\value{
returns a list with 2 items - \code{summary} and \code{details}.

\code{summary} is a dataframe with 5 columns, as described below:
\itemize{
\item {..name dependent on input..} =  named the same as \code{agg.poly.field} and
contains all of the values of \code{agg.poly.field}
\item \code{MARFIS_TRIPS} {This is how many Commercial trips had their majority of
their sets within this polygon}
\item \code{ISDB_TRIPS} {This is how many opbserved trips had their majority of
their sets within this polygon}
\item \code{MARFIS_SETS} {This is how many Commercial sets occurred in this polygon}
\item \code{ISDB_TRIPS} {This is how many Observed sets occurred in this polygon}
}

\code{details} is a list containing 4 dataframes - MARFIS_TRIPS, ISDB_TRIPS,
MARFIS_SETS and ISDB_SETS.  Each of these dataframes has 2 columns - the first
identifying the TRIP or SET, and the second identifying which area was
applied to that TRIP OR SET.
}
\description{
This function takes the results from get_marfis() and get_isdb()
and produces a table showing the how the proportion of observed data varies
across areas.  Different polygons can be provided, and the results will be
tailored to those polygons.
}
\note{
\itemize{
\item \code{_SETS} A given "TRIP_ID" in MARFIS or ISDB
typically has multiple sets, and each set has its own coordinates.  For this
analysis, every set is analyzed to see which polygon it occurred in, and
this information is provided via as the values for \code{MARFIS_SETS} and
\code{ISDB_SETS} in the results.

Coordinates for MARFIS_* data comes from the db table "PRO_SPC_INFO", and
coordinates for ISDB_*
 data come from "ISSETPROFILE" - using the first non-null
pair found while checking from P1-P4 in order.

\item \code{_TRIPS} = Since a single trip can have sets across
multiple polygons, this script determines which polygon had the most sets for
a given trip, and defines that polygon as the value for a given trip.  These
are provided as the values for \code{MARFIS_TRIPS} and \code{ISDB_TRIPS}.

\item Bad\code{Weird} Coordinates = If the latitude or longitude of the input data
for any set has a NULL value, an entry of "Bad coordinate" will be added to
the results.
Additionally, if fishing can occur outside of the extent of the provided polygon,
those trips/sets will be attributed to "Other".
}
}
\seealso{
Other simpleproducts: 
\code{\link{get_bycatch}()},
\code{\link{get_vmstracks}()},
\code{\link{plot_bycatch}()},
\code{\link{plot_coverage_time}()}

Other simpleproducts: 
\code{\link{get_bycatch}()},
\code{\link{get_vmstracks}()},
\code{\link{plot_bycatch}()},
\code{\link{plot_coverage_time}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
