% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_ISDB_Trip}
\alias{clean_ISDB_Trip}
\title{clean_ISDB_Trip}
\usage{
clean_ISDB_Trip(df = NULL, field = "ISDB_TRIP", out_name = "ISDB_TRIP_CLN")
}
\arguments{
\item{df}{default is \code{NULL}. This is a dataframe that includes a column of raw ISDB trip names}

\item{field}{default is \code{"ISDB_TRIP"}. This is the name of the field that contains the raw ISDB trip names}

\item{out_name}{default is \code{'ISDB_TRIP_CLN'}. The is the name of the new field that will be created i \code{df} after cleaning}
}
\value{
returns a list with 2 dataframes - "trips", and "sets".
"trips" contains all of the information necessary for identifying a trip
within MARFIS, as well as associated information about the trip
from the HAIL_OUTS and HAIL_IN_CALLS tables (e.g. confirmation numbers).
"sets" contains information about individual fishing activities, including
locations, dates, durations, gear amount, etc..
}
\description{
This function cleans ISDB trip names of spaces/odd characters to improve the potential for matching.
}
\seealso{
Other utils: 
\code{\link{paramOK}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{utils}
