% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enable_local}
\alias{enable_local}
\title{enable_local}
\usage{
enable_local(cxn = NULL, force.extract = FALSE, ...)
}
\arguments{
\item{cxn}{A valid Oracle connection object. This parameter allows you to
pass an existing connection, reducing the need to establish a new connection
within the function. If provided, it takes precedence over the connection-
related parameters.}

\item{force.extract}{default is \code{FALSE}  This flag forces a re-extraction of all of the
tables (rather than loading previously extracted versions from data.dir)}

\item{...}{other arguments passed to methods}
}
\description{
This function extracts all of the necessary oracle tables to a local folder so that
the functions can be run locally.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{setup}
