% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_.R
\name{fleet_}
\alias{fleet_}
\title{fleet_}
\usage{
fleet_(fleet = NULL, area = NULL, gearSpecs = NULL, ...)
}
\arguments{
\item{fleet}{This is an identifier that will be used to interrogate the licCore object to select the correct type(s), subtype(s), gear(s), and licence_species for the selected fleet}

\item{area}{This is an identifier that will be used to interrogate the licAreas object to select any NAFO areas associated with the selected fleet}

\item{gearSpecs}{This is an identifier that will be used to interrogate the licGearSpecs object to select any gear parameters associated with the selected fleet}

\item{...}{
  Arguments passed on to \code{\link[=set_defaults]{set_defaults}}
  \describe{
    \item{\code{dateStart}}{default is \code{NULL} format is \code{YYYYMMDD}, e.g. \code{dateStart = "20190219"}}
    \item{\code{dateEnd}}{default is \code{NULL} format is \code{YYYYMMDD}, e.g. \code{dateEnd = "20190219"}}
    \item{\code{year}}{default is \code{NULL} format is \code{YYYY}, e.g. \code{year = 2018}}
    \item{\code{lics}}{default is an empty dataframe\code{'all'}}
    \item{\code{marfSpp}}{default is \code{'all'} The marfis species code, usually sent by the fleet wrapper}
    \item{\code{useDate}}{default is \code{'LANDED_DATE'} Which ISDB date should be used while filtering ISDB data? "DATE_FISHED", or "LANDED_DATE"}
    \item{\code{returnFleet}}{default is \code{TRUE} Do you want a list object containing fleet information as part of your results?}
    \item{\code{returnMARFIS}}{default is \code{TRUE} Do you want a list object containing marfis trip and set information as part of your results? (requires returnFleet==T)}
    \item{\code{returnISDB}}{default is \code{TRUE} Do you want a list object containing isdb trip and set information as part of your results? (requires returnFleet==T & returnMARFIS ==T)}
    \item{\code{returnBycatch}}{default is \code{TRUE} Do you want a dataframe of all of the species caught during ISDB sets as part of your results?}
    \item{\code{returnLocations}}{default is \code{TRUE} Do you want a dataframe of the locations of the various sets as part of your results?}
    \item{\code{useReportedNAFO}}{default is \code{TRUE} (NOT YET IMPLEMENTED)}
    \item{\code{manual_fleet}}{default is \code{FALSE}}
    \item{\code{areas}}{default is \code{'NAFOSubunits_sf'}}
    \item{\code{areasField}}{default is \code{'NAFO_1'}}
    \item{\code{keepSurveyTrips}}{default is \code{FALSE}}
    \item{\code{matchMarfis}}{default is \code{TRUE}}
    \item{\code{matchMaxDayDiff}}{default is \code{15}}
    \item{\code{dropUnmatchedISDB}}{default is \code{TRUE}}
    \item{\code{useLocal}}{default is \code{FALSE}}
    \item{\code{data.dir}}{default is \code{'file.path(getwd(), "data")'} Necessary for useLocal == T.  This is the path to a folder where your *.rdata files are stored.}
    \item{\code{oracle.username}}{default is \code{'_none_'} This is your username for accessing oracle objects.}
    \item{\code{oracle.password}}{default is \code{'_none_'} This is your password for accessing oracle objects.}
    \item{\code{oracle.dsn}}{default is \code{'_none_'}  This is your dsn/ODBC identifier for accessing oracle objects.  Normally, the value should be "PTRAN"}
    \item{\code{usepkg}}{default is \code{'rodbc'} This indicates whether the connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc can
be slightly easier to setup, but roracle will extract data faster.}
    \item{\code{quietly}}{default is \code{TRUE}}
    \item{\code{debugISDBTrips}}{default is \code{'_none_'}}
    \item{\code{HS}}{default is \code{FALSE}}
    \item{\code{debuggit}}{default is \code{FALSE}}
    \item{\code{debugLics}}{default is \code{FALSE}}
  }}
}
\value{
list of objects, including marfis data, isdb data, information for matching isdb
and marfis data, and a summary of bycatch
}
\description{
This function is a generic wrapper that facilitates extracting fleet specific information using the data provided by fleet-specific wrappers.
}
\note{
Hardcoded parameters for this fleet are as follows:
\itemize{
  \item \code{marfSpp} = 608
}
}
\examples{
\dontrun{
test <- fleet_(fleet="POLLOCK_MOBILE", area="WESTERN", gearSpecs="SMALL" )
 }
}
\seealso{
Other fleets: 
\code{\link{fleet_haddock}()},
\code{\link{fleet_halibut}()},
\code{\link{fleet_pollock}()},
\code{\link{fleet_redfish}()},
\code{\link{fleet_scallop}()},
\code{\link{fleet_silverhake}()},
\code{\link{fleet_snowcrab}()},
\code{\link{fleet_surfclam}()},
\code{\link{fleet_winterflounder}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
