% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_halibut.R
\name{fleet_halibut}
\alias{fleet_halibut}
\title{fleet_halibut}
\usage{
fleet_halibut(vessLen = NULL, useLocal = NULL, ...)
}
\arguments{
\item{vessLen}{default is \code{NULL}.  Valid selections are pending.  NULL returns all vessels, "45LESS" will return vessels which their MARFIS licences
indicate they are less than 45 feet.  Licenced lengths in MARFIS can be different than what is stored in ISDB.}

\item{useLocal}{default is \code{NULL}.  Valid selections are  \code{TRUE} and \code{FALSE}

if \code{TRUE}, then the following parameter is necessary:
' \itemize{
  \item \code{data.dir} - This is a path on your computer to where your .RData files are stored.
  }

if \code{FALSE}, then the following parameters are necessary:
' \itemize{
  \item \code{oracle.username} - your existing oracle login name
  \item \code{oracle.password} - your existing oracle password
  \item \code{oracle.dsn} - usually "PTRAN" - the name of the database you're connecting to
  \item \code{usepkg} - default is \code{"rodbc"}, but \code{"roracle"} is also valid
  }}

\item{...}{
  Arguments passed on to \code{\link[=set_defaults]{set_defaults}}
  \describe{
    \item{\code{year}}{default is \code{NULL}.  \code{year} can be used if data should be extracted for an entire calendar year (i.e. Jan 1 --> Dec 31).  The format is
\code{YYYY}, e.g. \code{year = 2018}}
    \item{\code{dateStart}}{default is \code{NULL}.  This identifies the start date of the search window, in the format \code{YYYYMMDD} (e.g. \code{dateStart = "20190219"}).
If no \code{dateEnd} is provided, the window will be for 1 year (i.e, 365 days).  ,}
    \item{\code{dateEnd}}{default is \code{NULL} format is \code{YYYYMMDD}, e.g. \code{dateEnd = "20190219"}. \code{dateEnd} must be associated with a valid entry of
\code{dateStart} to identify a block of time for a data extraction  (e.g. Jun18, 2018 --> August 27, 2019).}
    \item{\code{marfSpp}}{default is \code{'all'}. The marfis species code, usually sent by the fleet wrapper}
    \item{\code{useDate}}{default is \code{'LANDED_DATE'}. Which ISDB date should be used while filtering ISDB data? "DATE_FISHED", or "LANDED_DATE"?}
    \item{\code{returnFleet}}{default is \code{TRUE}. Do you want a list object containing fleet information as part of your results?}
    \item{\code{returnMARFIS}}{default is \code{TRUE}. Do you want a list object containing marfis trip and set information as part of your results? (requires returnFleet==T)}
    \item{\code{returnISDB}}{default is \code{TRUE}. Do you want a list object containing isdb trip and set information as part of your results? (requires returnFleet==T & returnMARFIS ==T)}
    \item{\code{returnBycatch}}{default is \code{TRUE}. Do you want a dataframe of all of the species caught during ISDB sets as part of your results?}
    \item{\code{returnLocations}}{default is \code{TRUE}. Do you want a dataframe of the locations of the various sets as part of your results?}
    \item{\code{useReportedNAFO}}{default is \code{TRUE} (NOT YET IMPLEMENTED)}
    \item{\code{manual_fleet}}{default is \code{FALSE}.}
    \item{\code{areaFile}}{default is \code{'NAFOSubunits_sf'}.  This is used to identify which areas to check the trips and sets against. By default,
Mar.data::NAFOSubunits_sf is ued, but any objects in Mar.data could be used.}
    \item{\code{areaFileField}}{default is \code{'NAFO_1'}. This is a field within the \code{areas} object which specifies exactly which field of the areas object data
should be compared against.}
    \item{\code{keepSurveyTrips}}{default is \code{FALSE}. Within the ISDB database are non-commercial, survey trips.  Setting this to \code{TRUE}
ensures these trips are retained.}
    \item{\code{matchMarfis}}{default is \code{TRUE}. This indicates whether or not calls to get_ISDB should attempt to match MARFIS data.}
    \item{\code{matchMaxDayDiff}}{default is \code{15}. Any MARFIS and ISDB trips that vary by more than the # of days specified here will NOT be considered
matches (on the basis of common Vessel, licence and date).  They may still match on confirmation codes and/or trip names.}
    \item{\code{dropUnmatchedISDB}}{default is \code{TRUE}.}
    \item{\code{data.dir}}{default is \code{'file.path(getwd(), "data")'}. Necessary for useLocal == T.  This is the path to a folder where your *.rdata files are stored.}
    \item{\code{oracle.username}}{default is \code{'_none_'}. This is your username for accessing oracle objects.}
    \item{\code{oracle.password}}{default is \code{'_none_'}. This is your password for accessing oracle objects.}
    \item{\code{oracle.dsn}}{default is \code{'_none_'}.  This is your dsn/ODBC identifier for accessing oracle objects.  Normally, the value should be "PTRAN"}
    \item{\code{usepkg}}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc can
be slightly easier to setup, but roracle will extract data faster.}
    \item{\code{quietly}}{default is \code{TRUE}. This specifies whether or not status messages should be output to the console while the scripts run.}
    \item{\code{HS}}{default is \code{FALSE}. Setting this parameter to TRUE causes the package to try to imitate historic matching techniques.}
    \item{\code{debuggit}}{default is \code{FALSE}. If TRUE, this parameter causes the package to run in debug mode, providing much extraneous information.}
    \item{\code{debugLics}}{default is \code{NULL}.  If a vector of LICENCE_IDs is provided, the script will provide information about when the script drops them from
consideration.}
    \item{\code{debugVRs}}{default is \code{NULL}.  If a vector of VR numbers is provided, the script will provide information about when the script drops them from
consideration.}
    \item{\code{debugMARFTripIDs}}{default is \code{NULL}.  If a vector of MARFIS trip IDs is provided, the script will provide information about when the script drops them from
consideration.}
    \item{\code{debugISDBTripIDs}}{default is \code{NULL}.  If a vector of ISDB trip IDs is provided, the script will provide information about when the script drops them from
consideration.  Trip "names" are typically in a format like "J18-1234" or "J18-1234A".s}
    \item{\code{debugISDBTripNames}}{default is \code{NULL}.  If a vector of ISDB trip names is provided, the script will provide information about when the script drops them from
consideration.}
  }}
}
\value{
specific returned objects can be specified by the user, but the default result is a list of objects, including marfis data, isdb data, information for matching isdb
and marfis data, and a summary of bycatch, specifically:
a list item containing:
\itemize{
  \item \code{FLEET} - This is a dataframe of the unique combinations of (MARFIS) LICENCE_ID, VR_NUMBER and GEAR_CODE that
  was found for this fleet during the specified period
  \item \code{FLEET_ACTIVITY} - This is a dataframe of identifiers for all of the (MARFIS) fishing activity undertaken
  by vessels of this fleet during the specified period (i.e. LICENCE_ID, PRO_SPC_INFO_ID, LOG_EFRT_STD_INFO_ID, GEAR_CODE,
  MON_DOC_ID, VR_NUMBER, and several dates associated with the trip)
  \item \code{marf} - This is a list of 3 sets of information for the commercial catch data (i.e. marfis)-
  the trips, the sets, and a special dataframe containing information that can be used to link
  the commercial data to the ISDB data
  \item \code{isdb} - This is a list of 4 data objects - 2 of which are all of the discovered ISDB data
  TRIPS and SETS for the fleet, as well as the TRIPS and SETS from the observer data that were
  sucessfully matched with the MARFIS data
  \item \code{bycatch} - This is a dataframe with the various species that were observed during observed
  trips.  For each species, the estimated number caught, the estimated kept wt (kgs) and the
  estimated discarded wt(kg) are all captured
  }
}
\description{
This is a wrapper function that facilitates extracting information for the halibut fleet.
All of the information used to identify fleets is stored in the package's associated data files - licCore, licAreas,
and licGearSpecs.  The various wrappers can have different options (e.g. mobile vs fixed, western
vs eastern, 4XY vs 5ZJM, small mesh vs large mesh, diamond vs square mesh, etc), and depending on which options are selected,
different fleets are identified, and their data is extracted.
}
\note{
Hardcoded parameters for this fleet are as follows:
\itemize{
  \item \code{marfSpp} = 130
  \item \code{nafoCode} = c('3N\%','3O\%','3PS\%','4V\%','4W\%','4X\%','5\%')
  \item \code{gearCode} = c(51)
  \item \code{marfSpp} = 130
}
}
\examples{
\dontrun{
db <- fleet_halibut(vessLen = NULL,
                    useLocal = F,
                    year = 2018,
                    oracle.username = "<name>",
                    oracle.password="<password>",
                    oracle.dsn="PTRAN",
                    usepkg = "roracle"
                    )
local <- fleet_halibut(vessLen = "45LESS",
                       year = 2018,
                       useLocal = T,
                       data.dir = "c:/data_folder"
                      )
                       }
}
\seealso{
Other fleets: 
\code{\link{fleet_haddock}()},
\code{\link{fleet_pollock}()},
\code{\link{fleet_redfish}()},
\code{\link{fleet_scallop}()},
\code{\link{fleet_silverhake}()},
\code{\link{fleet_snowcrab}()},
\code{\link{fleet_surfclam}()},
\code{\link{fleet_winterflounder}()},
\code{\link{fleet_}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
