% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEARS}
\alias{GEARS}
\title{GEARS Metadata}
\format{
rda
\describe{
\item{GEAR_CODE}{code identifying the type of gear}
\item{GEAR}{name of the gear}
}
}
\usage{
data(GEARS)
}
\description{
This is an rda file of a lookup table used to associate different MARFIS gear codes with name of the gear:
}
\keyword{datasets}
