% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEARS_ISDB}
\alias{GEARS_ISDB}
\title{GEARS_ISDB Metadata}
\format{
rda
\describe{
\item{GEARCD_ID}{code identifying the type of gear}
\item{DESCRIPTION}{name of the gear}
}
}
\usage{
data(GEARS_ISDB)
}
\description{
This is an rda file of a lookup table used to associate different ISDB gear codes with name of the gear:
}
\keyword{datasets}
