% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEARS_MARFIS}
\alias{GEARS_MARFIS}
\title{GEARS_MARFIS Metadata}
\format{
rda
\describe{
\item{GEAR_CODE}{code identifying the type of gear}
\item{GEAR}{name of the gear}
}
}
\usage{
data(GEARS_MARFIS)
}
\description{
This is an rda file of a lookup table used to associate different MARFIS gear codes with name of the gear:
}
\keyword{datasets}
