% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{get_all}
\alias{get_all}
\title{get_all}
\usage{
get_all(...)
}
\arguments{
\item{...}{other arguments passed to methods}
}
\value{
returns a list including the fleet, and the resultant information from MARFIS, ISDB, and the bycatch
}
\description{
This function serves as a wrapper for calling \code{get_fleet(...)},\code{get_marfis(...)}
and \code{get_bycatch(...)} individually.
}
\note{
This function can accept many parameters.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
