% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bycatch.R
\name{get_bycatch}
\alias{get_bycatch}
\title{get_bycatch}
\usage{
get_bycatch(get_obs = NULL, ...)
}
\arguments{
\item{get_obs}{default is \code{NULL} this is the results of the get_obs() function}

\item{...}{other arguments passed to methods}
}
\value{
returns a dataframe of the bycatch information for the requested species (using the requested)
marfis species code and supplie results from get_obs
}
\description{
This function takes the results for get_obs() and a marfis species code
and generates a dataframe of all of the bycatche species from the observer db
}
\note{
This function can accept many parameters.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
