% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fleet_local.r
\name{get_fleet_local}
\alias{get_fleet_local}
\title{get_fleet_local}
\usage{
get_fleet_local(...)
}
\arguments{
\item{...}{other arguments passed to methods}
}
\value{
returns a data.frame with 6 columns - "GEAR_CODE", "GEAR_DESC",
        "MD_CODE", "MD_DESC", "VR_NUMBER", "LICENCE_ID"
}
\description{
This function extracts all of the Vessel/Licence combinations
associated with a particular fleet for a particular date range.
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS_local}()},
\code{\link{get_MARFIS_remote}()},
\code{\link{get_OBS_local}()},
\code{\link{get_OBS_remote}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet_remote}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
