% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdb.R
\name{get_isdb}
\alias{get_isdb}
\title{get_isdb}
\usage{
get_isdb(
  thisFleet = NULL,
  get_marfis = NULL,
  keepSurveyTrips = NULL,
  dateStart = NULL,
  dateEnd = NULL,
  extract_user = NULL,
  extract_computer = NULL,
  ...
)
}
\arguments{
\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from \code{Mar.fleets::get_fleet()}}

\item{get_marfis}{default is \code{NULL}. This is the list output by the
\code{Mar.fleets::get_marfis()} function - it contains dataframes of both the
trip and set information from MARFIS related to the specified fleet}

\item{keepSurveyTrips}{default is \code{FALSE}.  This indicates whether you
want to retain trips that were part of a survey.  These are not typical
commercial trips and their distribution will not reflect normal
commercial fishing patterns.}

\item{dateStart}{default is \code{NULL}. This is the start date (YYYY-MM-DD)
of the window of time you want to look at.}

\item{dateEnd}{default is \code{NULL}. This is the end date (YYYY-MM-DD)
of the window of time you want to look at.  If this is left blank, 1 year of
data will be returned.}

\item{extract_user}{default is \code{NULL}.  This parameter can be used with
\code{extract_computer} to load encypted data files extracted by another user
and/or computer}

\item{extract_computer}{default is \code{NULL}.  This parameter can be used with
\code{extract_user} to load encypted data files extracted by another user
and/or computer}

\item{...}{other arguments passed to methods}
}
\value{
returns a list with 2 dataframes - "ISDB_TRIPS", and "ISDB_SETS".
"ISDB_TRIPS" contains information for the ISDB trips, while "ISDB_SETS" contains information
about the ISDB sets.
}
\description{
This function extracts all of the records from the ISDB database that are
within a particular date range, and if a fleet is provided, it will also limit the results to
those vessels with particular combinations of VR and licence.
}
\seealso{
Other coreFuncs: 
\code{\link{fleet_}()},
\code{\link{getED}()},
\code{\link{getHIC}()},
\code{\link{getHOC}()},
\code{\link{getPS}()},
\code{\link{get_fleet}()},
\code{\link{get_isdb_sets}()},
\code{\link{get_isdb_trips}()},
\code{\link{get_marfis}()},
\code{\link{get_marfis_sets}()},
\code{\link{set_defaults}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
