% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isdb.R
\name{get_isdb_trips}
\alias{get_isdb_trips}
\title{get_isdb_trips}
\usage{
get_isdb_trips(mVR_LIC = NULL, ...)
}
\arguments{
\item{mVR_LIC}{default is \code{NULL}.}

\item{...}{other arguments passed to methods}
}
\value{
a dataframe
}
\description{
This function e
}
\seealso{
Other coreFuncs: 
\code{\link{fleet_}()},
\code{\link{getED}()},
\code{\link{getHIC}()},
\code{\link{getHOC}()},
\code{\link{getPS}()},
\code{\link{get_fleet}()},
\code{\link{get_isdb_sets}()},
\code{\link{get_isdb}()},
\code{\link{get_marfis_sets}()},
\code{\link{get_marfis}()},
\code{\link{set_defaults}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
