% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_marfis.R
\name{get_marfis}
\alias{get_marfis}
\title{get_marfis}
\usage{
get_marfis(thisFleet = NULL, ...)
}
\arguments{
\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from
\code{Mar.bycatch::get_fleet()}}

\item{...}{other arguments passed to methods}
}
\value{
returns a list with 2 dataframes - "trips", and "sets".
"trips" conta ins all of the information necessary for identifying a trip
within MARFIS, as well as associated information about the trip
from the HAIL_OUTS and HAIL_IN_CALLS tables (e.g. confirmation numbers).
"sets" contains information about individual fishing activities, including
locations, dates, durations, gear amount, etc..
}
\description{
This function extracts all of the MARFIS records for vessels with
particular combinations of VR_NUMBER and LICENCE_ID  for a given date range.
}
\seealso{
Other coreFuncs: 
\code{\link{get_fleet}()},
\code{\link{get_obs}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
