% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MARFIS_local.R
\name{get_MARFIS_local}
\alias{get_MARFIS_local}
\title{get_MARFIS_local}
\usage{
get_MARFIS_local(thisFleet = NULL, ...)
}
\arguments{
\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from
\code{Mar.bycatch::get_fleet()}}

\item{...}{other arguments passed to methods}
}
\value{
returns a list with 2 dataframes - "trips", and "sets".
"trips" conta ins all of the information necessary for identifying a trip
within MARFIS, as well as associated information about the trip
from the HAIL_OUTS and HAIL_IN_CALLS tables (e.g. confirmation numbers).
"sets" contains information about individual fishing activities, including
locations, dates, durations, gear amount, etc..
}
\description{
This function extracts all of the MARFIS records for vessels with
particular combinations of VR_NUMBER and LICENCE_ID  for a given date range.
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS_remote}()},
\code{\link{get_OBS_local}()},
\code{\link{get_OBS_remote}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet_local}()},
\code{\link{get_fleet_remote}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
