% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MARFIS_remote.R
\name{get_MARFIS_remote}
\alias{get_MARFIS_remote}
\title{get_MARFIS_remote}
\usage{
get_MARFIS_remote(
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  usepkg = "rodbc",
  data.dir = NULL,
  dateStart = NULL,
  dateEnd = NULL,
  thisFleet = NULL,
  useDate = "fished",
  marfSpp = NULL,
  nafoCode = NULL,
  vessLen = NULL,
  quietly = FALSE
)
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file),
this can be left and that value will be used.  If a value for this is
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{dateStart}{default is \code{NULL}. This is the start date (YYYY-MM-DD)
of the window of time you want to look at.}

\item{dateEnd}{default is \code{NULL}. This is the end date (YYYY-MM-DD)
of the window of time you want to look at.  If this is left blank, 1 year of
data will be returned.}

\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from
\code{Mar.bycatch::get_fleet()}}

\item{useDate}{default is \code{"fished"}. Some MARFIS tables have 2 different dates
that are used for recording when fishing activity took place.  One is "DATE_FISHED",
and the other is "LANDED_DATE". If useDate = "fished", the DATE_FISHED field will be used for
subsetting data by date.  Any other value will result in the use of "LANDED_DATE" instead.}

\item{marfSpp}{default is \code{NULL}.  This is the marfis species code for the species you want
records for. There are literally hundreds of codes, but here are some of the more commonly used:
\itemize{
\item 100 = Cod
\item 110 = Haddock
\item 120 = Redfish
\item 130 = Halibut
\item 141 = Yellowtail Flounder
\item 142 = Witch Flounder
\item 143 = Winter Flounder
\item 144 = Turbot
\item 170 = Pollock
\item 172 = Silver Hake
\item 200 = Herring
\item 251 = Swordfish
\item 608 = Surf Clam
\item 619 = Sea Cucumber
\item 623 = Sea Scallop
\item 700 = Lobster
}}

\item{nafoCode}{default is \code{NULL}.  This}

\item{vessLen}{default is \code{NULL}}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}
}
\value{
returns a list with 2 dataframes - "trips", and "sets".
"trips" conta ins all of the information necessary for identifying a trip
within MARFIS, as well as associated information about the trip
from the HAIL_OUTS and HAIL_IN_CALLS tables (e.g. confirmation numbers).
"sets" contains information about individual fishing activities, including
locations, dates, durations, gear amount, etc..
}
\description{
This function extracts all of the MARFIS records for vessels with
particular combinations of VR_NUMBER and LICENCE_ID  for a given date range.
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS_local}()},
\code{\link{get_MARFIS}()},
\code{\link{get_OBS_local}()},
\code{\link{get_OBS_remote}()},
\code{\link{get_OBS}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet_local}()},
\code{\link{get_fleet_remote}()},
\code{\link{get_fleet}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
