% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OBS.R
\name{get_OBS}
\alias{get_OBS}
\title{get_OBS}
\usage{
get_OBS(fn.oracle.username = "_none_", fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_", usepkg = "rodbc", dateStart = NULL,
  dateEnd = NULL, keepSurveyTrips = FALSE, quietly = FALSE,
  thisFleet = NULL)
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file),
this can be left and that value will be used.  If a value for this is
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{dateStart}{default is \code{NULL}. This is the start date (YYYY-MM-DD)
of the window of time you want to look at.}

\item{dateEnd}{default is \code{NULL}. This is the end date (YYYY-MM-DD)
of the window of time you want to look at.  If this is left blank, 1 year of
data will be returned.}

\item{keepSurveyTrips}{default is \code{FALSE}.  This indicates whether you
want to retain trips that were part of a survey.  These are not typical
commercial trips and their distribution will not reflect normal
commercial fishing patterns.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}

\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from
\code{Mar.bycatch::get_fleet()}}
}
\value{
returns a list with 4 dataframes - "MAP_OBS_MARFIS_TRIPS",
"MAP_OBS_MARFIS_SETS", "OBS_TRIPS", and "OBS_SETS".
"OBS_TRIPS" contains all of the information from OBS about the trip, but also
includes additional columns for the associated record in MARFIS. One column
("MATCHED_ON") identifies how marfis trips were matched to Observer trips.
Valid values for this field include any combination of:
\itemize{
\item OBS_TRIP - this means that both databases had the same entry for OBS_TRIP
(e.g. "J15-0001")
\item CONF_HI - this means that the observer database value for Logbook Confirmation
Number matched the value of HAIL_IN_CALLS.CONF_NUMBER.
\item CONF_HO - this means that the observer database value for Logbook Confirmation
Number matched the value of HAIL_OUTS.CONF_NUMBER.
\item VR_LIC_DATE - this means that the observer database had a record for a
specific combination
}
"OBS_SETS" contains all of the information from OBS about the sets, and also
includes additional columns for the associated sets in MARFIS.

from the HAIL_OUTS and HAIL_IN_CALLS tables (e.g. confirmation numbers).
"sets" contains information about individual fishing activities, including
locations, dates, durations, gear amount, etc..
}
\description{
This function takes the results from get_MARFIS() and extracts
all of the records from the observer database that can be matched using:
\itemize{
\item 1 = MARFIS confirmation numbers
\item 2 = Observer TRIP names (e.g. J18-0000)*
\item 3 = Correct combination of VRN and LICENCE and appropriate date range
}
* - only the alphanumeric characters of the trip names are used (e.g.
"J18-0000B" becomes "J180000").
}
\seealso{
Other fleets: \code{\link{calc_Coverage}},
  \code{\link{get_MARFIS}}, \code{\link{get_fleet}},
  \code{\link{match_sets}}, \code{\link{match_trips}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
 @export
}
\concept{fleets}
