% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OBS_remote.R
\name{get_OBS_remote}
\alias{get_OBS_remote}
\title{get_OBS_remote}
\usage{
get_OBS_remote(
  dateStart = NULL,
  dateEnd = NULL,
  keepSurveyTrips = FALSE,
  useDate = "fished",
  quietly = FALSE,
  thisFleet = NULL,
  get_MARFIS = NULL,
  ...
)
}
\arguments{
\item{dateStart}{default is \code{NULL}. This is the start date (YYYY-MM-DD)
of the window of time you want to look at.}

\item{dateEnd}{default is \code{NULL}. This is the end date (YYYY-MM-DD)
of the window of time you want to look at.  If this is left blank, 1 year of
data will be returned.}

\item{keepSurveyTrips}{default is \code{FALSE}.  This indicates whether you
want to retain trips that were part of a survey.  These are not typical
commercial trips and their distribution will not reflect normal
commercial fishing patterns.}

\item{useDate}{default is \code{"fished"}. Some MARFIS tables have 2 different dates
that are used for recording when fishing activity took place.  One is "DATE_FISHED",
and the other is "LANDED_DATE". If useDate = "fished", the DATE_FISHED field will be used for
subsetting data by date.  Any other value will result in the use of "LANDED_DATE" instead.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}

\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from
\code{Mar.bycatch::get_fleet()}}

\item{get_MARFIS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_MARFIS()} function - it contains dataframes of both the
trip and set information from MARFIS related to the specified fleet}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file),
this can be left and that value will be used.  If a value for this is
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}
}
\value{
returns a list with 2 dataframes - "OBS_TRIPS", and "OBS_SETS".
"OBS_TRIPS" contains information for the observer trips, while "OBS_SETS" contains information
about the observer sets.
}
\description{
This function extracts all of the records from the observer database that are
within a particular date range, and if a fleet is provided, it will also limit the results to
those vessels with particular combinations of VR and licence.
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS_local}()},
\code{\link{get_MARFIS_remote}()},
\code{\link{get_OBS_local}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet_local}()},
\code{\link{get_fleet_remote}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
