% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vmstracks.R
\name{get_vmstracks}
\alias{get_vmstracks}
\title{title get_vmstracks}
\usage{
get_vmstracks(get_marfis = NULL, get_obs = NULL, ...)
}
\arguments{
\item{get_marfis}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_marfis()} function - it contains dataframes of both the
trip and set information from MARFIS}

\item{get_obs}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_obs()} function - it contains dataframes of both the
trip and set information from the observer database.}

\item{...}{other arguments passed to methods}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}
}
\value{
returns a dataframe of the VMS data.  The OBS field contains a value>0 if the trip was observed.
}
\description{
This function takes the results from get_marfis() and get_obs()
and extracts the relevant VMS tracks, and flags whether or not each was observed
}
\seealso{
Other fleets: 
\code{\link{calc_coverage}()},
\code{\link{get_fleet}()},
\code{\link{get_marfis}()},
\code{\link{get_obs}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
