% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfis.R
\name{getPS}
\alias{getPS}
\title{getPS}
\usage{
getPS(allProSpc = NULL, ...)
}
\arguments{
\item{allProSpc}{default is \code{NULL}.}

\item{...}{other arguments passed to methods}
}
\value{
a dataframe
}
\description{
This function e
}
\seealso{
Other coreFuncs: 
\code{\link{fleet_}()},
\code{\link{getED}()},
\code{\link{getHIC}()},
\code{\link{getHOC}()},
\code{\link{get_fleet}()},
\code{\link{get_isdb_sets}()},
\code{\link{get_isdb_trips}()},
\code{\link{get_isdb}()},
\code{\link{get_marfis_sets}()},
\code{\link{get_marfis}()},
\code{\link{set_defaults}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
