% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{go}
\alias{go}
\title{go}
\usage{
go()
}
\value{
returns TRUE or FALSe, and informative messages
}
\description{
This function prevents people from running the package until it's ready
}
\seealso{
Other utils: 
\code{\link{clean_ISDB_Trip}()},
\code{\link{paramOK}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{utils}
