% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isDraft}
\alias{isDraft}
\title{isDraft}
\usage{
isDraft()
}
\value{
returns NULL
}
\description{
This function ensures that users are informed that the analytic is draft, and will
only run with their acknowledgment
}
\seealso{
Other utils: 
\code{\link{clean_ISDB_Trip}()},
\code{\link{go}()},
\code{\link{paramOK}()},
\code{\link{valuesOK}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{utils}
