% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{licAreas}
\alias{licAreas}
\title{licAreas Metadata}
\format{
rda
\describe{
\item{FLEET}{name of fleet}
\item{FLEET_AREA_ID}{name of subset of fleet for these areas}
\item{AREA_TYPE}{category of area}
\item{AREA}{specific area}
}
}
\usage{
data(licAreas)
}
\description{
This is an rda file with codes for identifying participants in different fleets.  It includes the fields identified below:
}
\keyword{datasets}
