% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{licCore}
\alias{licCore}
\title{licCore Metadata}
\format{
rda
\describe{
\item{FLEET}{name of fleet}
\item{LIC_TYPE}{Licence Type ID}
\item{LIC_SUBTYPE}{Licence Subtype ID}
\item{LIC_SP}{Licence species ID}
\item{LIC_GEAR}{Licence gear ID}
}
}
\usage{
data(licCore)
}
\description{
This is an rda file with codes for identifying participants in different fleets.  It includes the fields identified below:
}
\keyword{datasets}
