% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licGearSizes.R
\docType{data}
\name{licGearSizes}
\alias{licGearSizes}
\title{licGearSizes Metadata}
\format{
rda
\describe{
\item{FLEET}{name of fleet}
\item{FLEET_GEARSIZE_ID}{name of subset of fleet with these gear sizes}
\item{MIN}{smallest size cutoff}
\item{MAX}{largest size cutoff}
}
}
\usage{
licGearSizes
}
\description{
This is an rda file with codes for identifying participants in different fleets.  It includes the fields identified below:
}
\keyword{datasets}
