% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{licGearSpecs}
\alias{licGearSpecs}
\title{licGearSpecs Metadata}
\format{
rda
\describe{
\item{FLEET}{name of fleet}
\item{FLEET_GEARSPECS_ID}{name of subset of fleet with these gear sizes}
\item{MIN}{smallest size cutoff}
\item{MAX}{largest size cutoff}
\item{TYPE}{this describes the allowable gear - e.g. "S" (square); "D" (diamond)}
}
}
\usage{
data(licGearSpecs)
}
\description{
This is an rda file with codes for identifying participants in different fleets.  It includes the fields identified below:
}
\keyword{datasets}
