% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paramOK}
\alias{paramOK}
\title{paramOK}
\usage{
paramOK(useLocal = NULL, p = NULL)
}
\arguments{
\item{useLocal}{default is \code{NULL}. This specifies whether to run the script against local data or against Oracle (requires network or VPN).}

\item{p}{default is \code{NULL}. This is a list of the parameters sent by the user}
}
\value{
returns TRUE or FALSe, and informative messages
}
\description{
This function looks for issues with sent parameters.
}
\seealso{
Other utils: 
\code{\link{clean_ISDB_Trip}()},
\code{\link{go}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{utils}
