% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bycatch.R
\name{plot_bycatch}
\alias{plot_bycatch}
\title{plot_bycatch}
\usage{
plot_bycatch(
  obsSpp = NULL,
  df = NULL,
  showXSpp = NULL,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{obsSpp}{This is the observer db species code ID for the main species.}

\item{df}{default is \code{NULL}. This is the \code{bycatch} object from the \code{get_bycatch_()}
function.  Alternative, a data.frame with the following codes would work as well - "SPEC" (the code),
EST_NUM_CAUGHT (numeric), EST_KEPT_WT (numeric), EST_DISCARD_WT (numeric) and COMMON (species
common name).}

\item{showXSpp}{default is \code{NULL}. This is the total number of species that will be shown
on the plots.  Some datasets have so many species, that it is impossible to see them all.  When
clipping the data to the top species, it is sorted such that those species with the largest combined
EST_KEPT_WT and EST_DISCARD_WT are retained.}

\item{title}{default is \code{NULL}. This will be the title of your plots.}

\item{subtitle}{default is \code{NULL}. This will be the subtitle of your plots.}
}
\value{
nothing, but png files will be generated in your working directory.
}
\description{
This function generates simple pdf plots for the output from Mar.bycatch species
functions.
}
\examples{
\dontrun{
plot_bycatch(obsSpp = swordfish2018$bycatch[1,1],
             df = swordfish2018$bycatch, showXSpp = 20,
             title ="Swordfish",
             subtitle =2018)
                                }
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
