% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults2.R
\name{set_defaults2}
\alias{set_defaults2}
\title{set_defaults2}
\usage{
set_defaults2(
  lics = "all",
  marfSpp = "all",
  gearCode = "all",
  nafoCode = "all",
  gearSpType = "all",
  gearSpSize = "all",
  vessLen = "all",
  useDate = "LANDED_DATE",
  fleetOnly = T,
  useReportedNAFO = TRUE,
  manual_fleet = F,
  areas = "NAFOSubunits_sf",
  areasField = "NAFO_1",
  dateStart = NULL,
  dateEnd = NULL,
  year = NULL,
  keepSurveyTrips = FALSE,
  matchMarfis = TRUE,
  matchMaxDayDiff = 15,
  dropUnmatchedISDB = TRUE,
  data.dir = file.path(getwd(), "data"),
  oracle.username = "_none_",
  oracle.password = "_none_",
  oracle.dsn = "_none_",
  usepkg = "rodbc",
  useLocal = FALSE,
  quietly = TRUE,
  debugISDBTrips = "_none_",
  HS = FALSE,
  debug = FALSE,
  ...
)
}
\arguments{
\item{lics}{default is \code{'all'}}

\item{marfSpp}{default is \code{'all'}}

\item{gearCode}{default is \code{'all'}}

\item{nafoCode}{default is \code{'all'}}

\item{gearSpType}{default is \code{'all'}}

\item{gearSpSize}{default is \code{'all'}}

\item{vessLen}{default is \code{'all'}}

\item{useDate}{default is \code{'LANDED_DATE'}}

\item{fleetOnly}{default is \code{TRUE}}

\item{useReportedNAFO}{default is \code{TRUE}}

\item{areas}{default is \code{'NAFOSubunits_sf'}}

\item{areasField}{default is \code{'NAFO_1'}}

\item{dateStart}{default is \code{NULL}}

\item{dateEnd}{default is \code{NULL}}

\item{year}{default is \code{NULL}}

\item{keepSurveyTrips}{default is \code{FALSE}}

\item{matchMarfis}{default is \code{TRUE}}

\item{matchMaxDayDiff}{default is \code{15}}

\item{dropUnmatchedISDB}{default is \code{TRUE}}

\item{data.dir}{default is \code{'file.path(getwd(), "data")'}}

\item{oracle.username}{default is \code{'_none_'}}

\item{oracle.password}{default is \code{'_none_'}}

\item{oracle.dsn}{default is \code{'_none_'}}

\item{usepkg}{default is \code{'rodbc'}}

\item{useLocal}{default is \code{FALSE}}

\item{debugISDBTrips}{default is \code{'_none_'}}

\item{HS}{default is \code{FALSE}}

\item{...}{other arguments passed to methods}

\item{manual_fleetdefault}{is \code{FALSE}}

\item{quietlydefault}{is \code{TRUE}}

\item{debugdefault}{is \code{FALSE}}
}
\description{
This function ensures that all possible parameters are provided.  They will be
overwritten by any sent directly to the fleet_<species> or get_all functions.
}
\seealso{
Other coreFuncs: 
\code{\link{get_fleet}()},
\code{\link{get_isdb}()},
\code{\link{get_marfis}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
