% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_halibut.R
\name{sp_halibut}
\alias{sp_halibut}
\title{sp_halibut}
\usage{
sp_halibut(year = NULL, vessLen = c(0, 999), ...)
}
\arguments{
\item{year}{default is \code{NULL}. This is a year (YYYY) for which you want to look at the marfis,
observer and bycatch data.}

\item{vessLen}{default is \code{NULL}.  This is a vector of vessel lengths.  If it is not NULL or
"all", it will be used to restrict vessels by their size.  The supplied vector will only be assessed
for its max and min values, so if you wanted vessels up to and including 45ft, you could enter either
of the following - \code{c(0,45)} or \code{seq(0,45,1)}.}

\item{...}{other arguments passed to methods}
}
\value{
list of objects, including marfis data, observer data, information for matching observer
and marfis data, and a summary of bycatch
}
\description{
This function is a wrapper function that facilitates extracting the following
information for the halibut fleets:
\itemize{
  \item \code{fleet} - This is a dataframe of identifiers for all of the various trips undertaken by the
  selected fleet for the specified period (e.g. VRNs, licence IDs, Monitoring Document #s, etc)
  \item \code{marf} - This is a list of 3 sets of information for the commercial catch data (i.e. marfis)-
  the trips, the sets, and a special dataframe containing information that can be used to link
  the commercial data to the observer data
  \item \code{obs} - This is a list of 4 data objects - 2 of which are all of the discovered observer data
  TRIPS and SETS for the fleet, as well as the TRIPS and SETS from teh observer data that were
  sucessfully matched with the MARFIS data
  \item \code{bycatch} - This is a dataframe with the various species that were observed during observed
  trips.  For each species, the estimated number caught, the estimated kept wt (kgs) and the
  estimated discarded wt(kg) are all captured
}
}
\examples{
\dontrun{
Halibut <- sp_halibut(year = 2018, vessLen = c(0,45), data.dir = "C:/myData")
}
}
\seealso{
Other species: 
\code{\link{sp_haddock}()},
\code{\link{sp_northernshrimp}()},
\code{\link{sp_offshorelobster}()},
\code{\link{sp_pollock}()},
\code{\link{sp_redfish}()},
\code{\link{sp_silverhake}()},
\code{\link{sp_snowcrab}()},
\code{\link{sp_surfclam}()},
\code{\link{sp_swordfish}()},
\code{\link{sp_winterflounder}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{species}
