% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_pollock.R
\name{sp_pollock}
\alias{sp_pollock}
\title{sp_pollock}
\usage{
sp_pollock(type = NULL, mesh = NULL, component = NULL, ...)
}
\arguments{
\item{type}{default is \code{NULL}. This is either "FIXED" or "MOBILE".}

\item{mesh}{default is \code{NULL}. This is either "SMALL" (i.e. 1-129mm) or "LARGE" (i.e. 130mm+), or "ALL".}

\item{component}{default is \code{NULL}. This is either "WESTERN" or "EASTERN".}

\item{...}{other arguments passed to methods}
}
\value{
list of objects, including marfis data, observer data, information for matching observer
and marfis data, and a summary of bycatch
}
\description{
This function is a wrapper function that facilitates extracting the following
information for the pollock fleets:
\itemize{
  \item \code{fleet} - This is a dataframe of identifiers for all of the various trips undertaken by the
  selected fleet for the specified period (e.g. VRNs, licence IDs, Monitoring Document #s, etc)
  \item \code{marf} - This is a list of 3 sets of information for the commercial catch data (i.e. marfis)-
  the trips, the sets, and a special dataframe containing information that can be used to link
  the commercial data to the observer data
  \item \code{obs} - This is a list of 4 data objects - 2 of which are all of the discovered observer data
  TRIPS and SETS for the fleet, as well as the TRIPS and SETS from teh observer data that were
  sucessfully matched with the MARFIS data
  \item \code{bycatch} - This is a dataframe with the various species that were observed during observed
  trips.  For each species, the estimated number caught, the estimated kept wt (kgs) and the
  estimated discarded wt(kg) are all captured
}
}
\examples{
\dontrun{
Pollock_West_m_sm <- sp_pollock(type = "MOBILE",
                                mesh="SMALL", component = "WESTERN",
                                data.dir = "C:/myData")
                                }
\dontrun{
Pollock_East_m_lg <- sp_pollock(type = "MOBILE",
                                mesh="LARGE", component = "EASTERN",
                                data.dir = "C:/myData")
                                }
}
\seealso{
Other species: 
\code{\link{sp_haddock}()},
\code{\link{sp_halibut}()},
\code{\link{sp_northernshrimp}()},
\code{\link{sp_offshorelobster}()},
\code{\link{sp_redfish}()},
\code{\link{sp_silverhake}()},
\code{\link{sp_snowcrab}()},
\code{\link{sp_surfclam}()},
\code{\link{sp_swordfish}()},
\code{\link{sp_winterflounder}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{species}
