% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_surfclam.R
\name{sp_surfclam}
\alias{sp_surfclam}
\title{sp_surfclam}
\usage{
sp_surfclam(...)
}
\arguments{
\item{...}{other arguments passed to methods}
}
\value{
list of objects, including marfis data, isdb data, information for matching isdb
and marfis data, and a summary of bycatch
}
\description{
This function is a wrapper function that facilitates extracting the following
information for the surfclam fleet:
\itemize{
  \item \code{fleet} - This is a dataframe of identifiers for all of the various trips undertaken by the
  selected fleet for the specified period (e.g. VRNs, licence IDs, Monitoring Document #s, etc)
  \item \code{marf} - This is a list of 3 sets of information for the commercial catch data (i.e. marfis)-
  the trips, the sets, and a special dataframe containing information that can be used to link
  the commercial data to the ISDB data
  \item \code{isdb} - This is a list of 4 data objects - 2 of which are all of the discovered ISDB data
  TRIPS and SETS for the fleet, as well as the TRIPS and SETS from the observer data that were
  sucessfully matched with the MARFIS data
  \item \code{bycatch} - This is a dataframe with the various species that were observed during observed
  trips.  For each species, the estimated number caught, the estimated kept wt (kgs) and the
  estimated discarded wt(kg) are all captured
  }
}
\examples{
\dontrun{
Surfclam <- sp_surfclam(data.dir = "C:/myData")
                       }
}
\seealso{
Other species: 
\code{\link{sp_haddock}()},
\code{\link{sp_halibut}()},
\code{\link{sp_northernshrimp}()},
\code{\link{sp_offshorelobster}()},
\code{\link{sp_pollock}()},
\code{\link{sp_redfish}()},
\code{\link{sp_silverhake}()},
\code{\link{sp_swordfish}()},
\code{\link{sp_winterflounder}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{species}
