% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_swordfish.R
\name{sp_swordfish}
\alias{sp_swordfish}
\title{sp_swordfish}
\usage{
sp_swordfish(data.dir = NULL, year = NULL)
}
\arguments{
\item{data.dir}{The default is your working directory. If you are hoping to
load existing data, this folder should identify the folder containing your
*.rdata files.}

\item{year}{default is \code{NULL}. This is a year (YYYY) for which you want to look at the marfis,
observer and bycatch data.}
}
\value{
list of objects, including marfis data, observer data, information for matching observer
and marfis data, and a summary of bycatch
}
\description{
This function is a wrapper function that facilitates extracting the following
information for the swordfish fleet:
\itemize{
  \item \code{fleet} - This is a dataframe of identifiers for all of the various trips undertaken by the
  selected fleet for the specified period (e.g. VRNs, licence IDs, Monitoring Document #s, etc)
  \item \code{marf} - This is a list of 3 sets of information for the commercial catch data (i.e. marfis)-
  the trips, the sets, and a special dataframe containing information that can be used to link
  the commercial data to the observer data
  \item \code{obs} - This is a list of 4 data objects - 2 of which are all of the discovered observer data
  TRIPS and SETS for the fleet, as well as the TRIPS and SETS from teh observer data that were
  sucessfully matched with the MARFIS data
  \item \code{bycatch} - This is a dataframe with the various species that were observed during observed
  trips.  For each species, the estimated number caught, the estimated kept wt (kgs) and the
  estimated discarded wt(kg) are all captured
}
}
\examples{
\dontrun{
Swordfish <- sp_swordfish(data.dir = "C:/myData",
                          year = 2018)
                          }
}
\seealso{
Other species: 
\code{\link{sp_haddock}()},
\code{\link{sp_halibut}()},
\code{\link{sp_offshore_lobster}()},
\code{\link{sp_pollock}()},
\code{\link{sp_redfish}()},
\code{\link{sp_silverhake}()},
\code{\link{sp_surfclam}()},
\code{\link{sp_winterflounder}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{species}
