% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPECIES_ISDB}
\alias{SPECIES_ISDB}
\title{SPECIES_ISDB Metadata}
\format{
rda
\describe{
\item{SPECCCD_ID}{code identifying the species}
\item{COMMON}{common name of the species}
\item{SCIENTIFIC}{scientific name of the species}
}
}
\usage{
data(SPECIES_ISDB)
}
\description{
This is an rda file of a lookup table used to associate different ISDB species codes with the name of the species:
}
\keyword{datasets}
