% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPECIES_MARFIS}
\alias{SPECIES_MARFIS}
\title{SPECIES_MARFIS Metadata}
\format{
rda
\describe{
\item{SPECIES_CODE}{code identifying the species}
\item{SPECIES_NAME}{name of the species/species group}
}
}
\usage{
data(SPECIES_MARFIS)
}
\description{
This is an rda file of a lookup table used to associate different MARFIS species codes with the name of the species:
}
\keyword{datasets}
