% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spLookups}
\alias{spLookups}
\title{spLookups Metadata}
\format{
rda
\describe{
\item{SPECCD_ID}{ISDB species code}
\item{SCIENTIFIC}{Scientifi name}
\item{COMMON}{Common name}
\item{MARFIS_CODE}{marfis species code}
}
}
\usage{
spLookups
}
\description{
This is an rda file with both observer and MARFIS species codes.  It is used to allow
identification of ISDB species, when only marfis species codes are required
}
\keyword{datasets}
