% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizer.R
\name{summarizer}
\alias{summarizer}
\title{summarizer}
\usage{
summarizer(
  data = NULL,
  units = "KGS",
  bySpp = TRUE,
  byYr = TRUE,
  byGr = TRUE,
  byNAFO = FALSE,
  byCust = NULL,
  speccd_id = NULL,
  doMARF = T,
  doISDB = T
)
}
\arguments{
\item{data}{default is \code{NULL}. This is the entire output from any of the fleet wrappers.}

\item{units}{default is \code{"KGS"}. This is the units you want weights returned in.  Valid
options are any of "KGS", "TONNES", "LBS".}

\item{bySpp}{default is \code{TRUE}. If TRUE, the summed weights and numbers of unique trips and
licence_ids will will be broken down by all identified landed species. If FALSE,
weights will be summed, irrespective of landed species.}

\item{byYr}{default is \code{TRUE}.   If TRUE, the summed weights and numbers of unique trips
and licence_ids will will be broken down by each year in which a landing was identified. If FALSE,
weights will be summed, irrespective of year. "Year" is taken from the end date of the associated
trip record.}

\item{byGr}{default is \code{TRUE}. If TRUE, the summed weights and numbers of unique trips and
licence_ids will will be broken down by all identified gear types. If FALSE,
weights will be summed, irrespective of gear.}

\item{byNAFO}{default is \code{FALSE}.   If TRUE, the summed weights and numbers of unique trips
and licence_ids will will be broken down by <reported> NAFO division in which a landing was
identified. If FALSE, weights will be summed, irrespective of <reported> NAFO division.}

\item{byCust}{default is \code{NULL}. This can be the name of any field found in \code{<data>$marf$MARF_SETS}
by which you would like to see aggregated results}

\item{doMARF}{default is \code{T}. This indicates that the MARFIS results should be summarized.}

\item{doISDB}{default is \code{T}. This indicates that the ISDB results should be summarized.}
}
\value{
a data frame
}
\description{
This is a simple tool that quickly breaks down the results of any of the fleet
wrappers to provide information on the number of trips, unique vessels, unique licences, and total
catch.  These results can be aggregated by YEAR, SPECIES_CODE, and GEAR TYPE.
}
\examples{
\dontrun{
summary <- summarizer(data = Halibut2017, byGr = TRUE)
       }
}
\seealso{
Other simpleproducts: 
\code{\link{get_vmstracks}()},
\code{\link{quickMap}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
