% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{valuesOK}
\alias{valuesOK}
\title{valuesOK}
\usage{
valuesOK(valSent = NULL, valOK = NULL, valID = "this vector")
}
\arguments{
\item{valSent}{default is \code{NULL}. This is a vector of values sent by the user}

\item{valOK}{default is \code{NULL}. This is a vector of of known, acceptable values}

\item{valID}{default is \code{"this vector"}. This is an identifier for the vector.}
}
\value{
returns NULL
}
\description{
This function ensures that the user has not tried to send unacceptable values
}
\seealso{
Other utils: 
\code{\link{clean_ISDB_Trip}()},
\code{\link{go}()},
\code{\link{isDraft}()},
\code{\link{paramOK}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{utils}
