% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Coverage.R
\name{calc_Coverage}
\alias{calc_Coverage}
\title{calc_Coverage}
\usage{
calc_Coverage(
  get_OBS = NULL,
  get_MARFIS = NULL,
  agg.poly.shp = NULL,
  agg.poly.field = NULL,
  quietly = FALSE
)
}
\arguments{
\item{get_OBS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_OBS()} function - it contains dataframes of both the
trip and set information from OBSERVER related to the specified fleet}

\item{get_MARFIS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_MARFIS()} function - it contains dataframes of both the
trip and set information from MARFIS related to the specified fleet}

\item{agg.poly.shp}{default is \code{NULL}.  This is the shapefile that has
polygons that should be checked for sufficient unique values of the
sens.fields.  If NULL, NAFO zones will be used.  Otherwise, a path to any
polygon shapefile can be provided.}

\item{agg.poly.field}{default is \code{NULL}.  This identifies the field within
the shapefile provided to agg.poly.shp that should be used to check for
sufficient unique values of the sens.fields.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}
}
\value{
returns a list with 2 items - \code{summary} and \code{details}.

\code{summary} is a dataframe with 5 columns, as described below:
\itemize{
\item {..name dependent on input..} =  named the same as \code{agg.poly.field} and
contains all of the values of \code{agg.poly.field}
\item \code{MARFIS_TRIPS} {This is how many Commercial trips had their majority of
their sets within this polygon}
\item \code{OBS_TRIPS} {This is how many opbserved trips had their majority of
their sets within this polygon}
\item \code{MARFIS_SETS} {This is how many Commercial sets occurred in this polygon}
\item \code{OBS_TRIPS} {This is how many Observed sets occurred in this polygon}
}

\code{details} is a list containing 4 dataframes - MARFIS_TRIPS, OBS_TRIPS,
MARFIS_SETS and OBS_SETS.  Each of these dataframes has 2 columns - the first
identifying the TRIP or SET, and the second identifying which area was
applied to that TRIP OR SET.
}
\description{
This function takes the results from get_MARFIS() and get_OBS()
and produces a table showing the how the proportion of observed data varies
across areas.  Different polygons can be provided, and the results will be
tailored to those polygons.
}
\note{
\itemize{
\item \code{_SETS} A given "TRIP_ID" in MARFIS or OBSERVER
typically has multiple sets, and each set has its own coordinates.  For this
analysis, every set is analyzed to see which polygon it occurred in, and
this information is provided via as the values for \code{MARFIS_SETS} and
\code{OBS_SETS} in the results.

Coordinates for MARFIS_* data comes from the db table "PRO_SPC_INFO", and
coordinates for OBS_* data come from "ISSETPROFILE" - using the first non-null
pair found while checking from P1-P4 in order.

\item \code{_TRIPS} = Since a single trip can have sets across
multiple polygons, this script determines which polygon had the most sets for
a given trip, and defines that polygon as the value for a given trip.  These
are provided as the values for \code{MARFIS_TRIPS} and \code{OBS_TRIPS}.

\item Bad\code{Weird} Coordinates = If the latitude or longitude of the input data
for any set has a NULL value, an entry of "Bad coordinate" will be added to
the results.
Additionally, if fishing can occur outside of the extent of the provided polygon,
those trips/sets will be attributed to "Other".
}
}
\seealso{
Other fleets: 
\code{\link{get_MARFIS_local}()},
\code{\link{get_MARFIS_remote}()},
\code{\link{get_OBS_local}()},
\code{\link{get_OBS_remote}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet_local}()},
\code{\link{get_fleet_remote}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
