% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_local.R
\name{enable_local}
\alias{enable_local}
\title{enable_local}
\usage{
enable_local(
  data.dir = NULL,
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  usepkg = "rodbc"
)
}
\arguments{
\item{data.dir}{The default is your working directory. If you are hoping to
load existing data, this folder should identify the folder containing your
*.rdata files.}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file),
this can be left and that value will be used.  If a value for this is
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}
}
\description{
This function extracts all of the necessary oracle tables to a local folder so that
the functions can be run locally.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{setup}
