% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_.R
\name{fleet_}
\alias{fleet_}
\title{fleet_}
\usage{
fleet_(fleet = NULL, area = NULL, gearSpecs = NULL, ...)
}
\arguments{
\item{fleet}{This is an identifier that will be used to interrogate the licCore object to select the correct type(s), subtype(s), gear(s), and licence_species for the selected fleet}

\item{area}{This is an identifier that will be used to interrogate the licAreas object to select any NAFO areas associated with the selected fleet}

\item{gearSpecs}{This is an identifier that will be used to interrogate the licGearSpecs object to select any gear parameters associated with the selected fleet}

\item{...}{
  Arguments passed on to \code{\link[=set_defaults]{set_defaults}}
  \describe{
    \item{\code{lics}}{default is an empty dataframe\code{'all'}}
    \item{\code{marfSpp}}{default is \code{'all'}}
    \item{\code{vessLen}}{default is \code{'all'}}
    \item{\code{useDate}}{default is \code{'LANDED_DATE'}}
    \item{\code{returnFleet}}{default is \code{TRUE}}
    \item{\code{returnMARFIS}}{default is \code{TRUE}}
    \item{\code{returnISDB}}{default is \code{TRUE}}
    \item{\code{returnBycatch}}{default is \code{TRUE}}
    \item{\code{returnLocations}}{default is \code{TRUE}}
    \item{\code{useReportedNAFO}}{default is \code{TRUE}}
    \item{\code{manual_fleet}}{default is \code{FALSE}}
    \item{\code{areas}}{default is \code{'NAFOSubunits_sf'}}
    \item{\code{areasField}}{default is \code{'NAFO_1'}}
    \item{\code{dateStart}}{default is \code{NULL}}
    \item{\code{dateEnd}}{default is \code{NULL}}
    \item{\code{year}}{default is \code{NULL}}
    \item{\code{keepSurveyTrips}}{default is \code{FALSE}}
    \item{\code{matchMarfis}}{default is \code{TRUE}}
    \item{\code{matchMaxDayDiff}}{default is \code{15}}
    \item{\code{dropUnmatchedISDB}}{default is \code{TRUE}}
    \item{\code{data.dir}}{default is \code{'file.path(getwd(), "data")'}}
    \item{\code{oracle.username}}{default is \code{'_none_'}}
    \item{\code{oracle.password}}{default is \code{'_none_'}}
    \item{\code{oracle.dsn}}{default is \code{'_none_'}}
    \item{\code{usepkg}}{default is \code{'rodbc'}}
    \item{\code{useLocal}}{default is \code{FALSE}}
    \item{\code{quietly}}{default is \code{TRUE}}
    \item{\code{debugISDBTrips}}{default is \code{'_none_'}}
    \item{\code{HS}}{default is \code{FALSE}}
  }}
}
\value{
list of objects, including marfis data, isdb data, information for matching isdb
and marfis data, and a summary of bycatch
}
\description{
This function is a generic wrapper that facilitates extracting fleet specific information using the data provided by fleet-specific wrappers.
}
\note{
Hardcoded parameters for this fleet are as follows:
\itemize{
  \item \code{marfSpp} = 608
}
}
\examples{
\dontrun{
test <- fleet_(fleet="POLLOCK_MOBILE", area="WESTERN", gearSpecs="SMALL" )
 }
}
\seealso{
Other fleets: 
\code{\link{fleet_haddock}()},
\code{\link{fleet_halibut}()},
\code{\link{fleet_pollock}()},
\code{\link{fleet_redfish}()},
\code{\link{fleet_scallop}()},
\code{\link{fleet_silverhake}()},
\code{\link{fleet_snowcrab}()},
\code{\link{fleet_surfclam}()},
\code{\link{fleet_winterflounder}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
