% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_.R
\name{fleet_}
\alias{fleet_}
\title{fleet_}
\usage{
fleet_(fleet = NULL, area = NULL, gearSpecs = NULL, ...)
}
\arguments{
\item{fleet}{This is an identifier that will be used to interrogate the licCore object to select the correct type(s), subtype(s), gear(s), and licence_species for the selected fleet}

\item{area}{This is an identifier that will be used to interrogate the licAreas object to select any NAFO areas associated with the selected fleet}

\item{gearSpecs}{This is an identifier that will be used to interrogate the licGearSpecs object to select any gear parameters associated with the selected fleet}

\item{...}{
  Arguments passed on to \code{\link[=set_defaults]{set_defaults}}
  \describe{
    \item{\code{year}}{default is \code{NULL}.  \code{year} can be used if data should be extracted for an entire calendar year (i.e. Jan 1 --> Dec 31).  The format is
\code{YYYY}, e.g. \code{year = 2018}.  \code{dateStart} takes priority over \code{year} if both are provided.}
    \item{\code{dateStart}}{default is \code{NULL}.  This identifies the start date of the search window, in the format \code{YYYY-MM-DD} (e.g. \code{dateStart = "2019-02-19"}).
If no \code{dateEnd} is provided, the window will be for 1 year (i.e, 365 days).  ,}
    \item{\code{dateEnd}}{default is \code{NULL} format is \code{YYYY-MM-DD}, e.g. \code{dateEnd = "2019-02-19"}. \code{dateEnd} must be associated with a valid entry of
\code{dateStart} to identify a block of time for a data extraction  (e.g. Jun18, 2018 --> August 27, 2019).}
    \item{\code{lics}}{default is \code{'all'}.  Fleet wrappers will populate it with a dataframe specifying the licence types, subtypes, gear and licence species valid
for a particular fleet.  This information comes from  \code{Mar.bycatch::licCore}.}
    \item{\code{marfSpp}}{default is \code{'all'}. The marfis species code, usually sent by the fleet wrapper}
    \item{\code{useDate}}{default is \code{'LANDED_DATE'}. Which ISDB date should be used while filtering ISDB data?  \code{DATE_FISHED}, or \code{LANDED_DATE}?}
    \item{\code{returnMARFIS}}{default is \code{TRUE}. Do you want a list object containing marfis trip and set information as part of your results? (requires \code{returnFleet = T})}
    \item{\code{returnISDB}}{default is \code{TRUE}. Do you want a list object containing isdb trip and set information as part of your results? (requires \code{returnFleet = T} & \code{returnMARFIS = T})}
    \item{\code{manual_fleet}}{default is \code{FALSE}.}
    \item{\code{areaFile}}{default is \code{'NAFOSubunits_sf'}.  This is used to identify which areas to check the trips and sets against. By default,
Mar.data::NAFOSubunits_sf is ued, but any objects in Mar.data could be used.}
    \item{\code{areaFileField}}{default is \code{'NAFO_1'}. This is a field within the \code{areas} object which specifies exactly which field of the areas object data
should be compared against.}
    \item{\code{nafoDet}}{default is \code{2}, but values between \code{1} and \code{4} are acceptable..  This specifies the level of detail that will be used in the
summarized locations table.  Using the default value of 2, trips and sets will be summarized by areas such as "4X", "4V" and "5Z" (i.e 2 characters).  If
set to "1", areas would be more general  (e.g. "3", "4", "5"; i.e. 1 character) , while a value like 4 would summarize the trips and sets into very specific
NAFO subunits (e.g. "3PSA","4VSB" and "5ZEM")}
    \item{\code{keepSurveyTrips}}{default is \code{FALSE}. Within the ISDB database are non-commercial, survey trips.  Setting this to \code{TRUE}
ensures these trips are retained.}
    \item{\code{matchMarfis}}{default is \code{TRUE}. This indicates whether or not calls to get_ISDB should attempt to match MARFIS data.}
    \item{\code{matchMaxDayDiff}}{default is \code{2}. Any MARFIS and ISDB trips that vary by more than the # of days specified here will NOT be considered
matches (on the basis of common Vessel, licence and date).  They may still match on confirmation codes and/or trip names.}
    \item{\code{dropUnmatchedISDB}}{default is \code{TRUE}.}
    \item{\code{useLocal}}{default is \code{FALSE}. TThis specifies whether to run the script against local data or against Oracle (requires network or VPN).}
    \item{\code{data.dir}}{default is \code{'file.path(getwd(), "data")'}. Necessary for useLocal == T.  This is the path to a folder where your *.rdata files are stored.}
    \item{\code{oracle.username}}{default is \code{'_none_'}. This is your username for accessing oracle objects.}
    \item{\code{oracle.password}}{default is \code{'_none_'}. This is your password for accessing oracle objects.}
    \item{\code{oracle.dsn}}{default is \code{'_none_'}.  This is your dsn/ODBC identifier for accessing oracle objects.  Normally, the value should be "PTRAN"}
    \item{\code{usepkg}}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc can
be slightly easier to setup, but roracle will extract data faster.}
    \item{\code{quietly}}{default is \code{TRUE}. This specifies whether or not status messages should be output to the console while the scripts run.#' @param HS default is \code{FALSE}. Setting this parameter to TRUE causes the package to try to imitate historic matching techniques.}
    \item{\code{HS}}{default is \code{FALSE}. Setting this parameter to TRUE causes the package to try to imitate historic matching techniques.}
    \item{\code{debuggit}}{default is \code{FALSE}. If TRUE, this parameter causes the package to run in debug mode, providing much extraneous information.}
    \item{\code{debugLics}}{default is \code{NULL}.  If a vector of LICENCE_IDs is provided, the script will provide information about when the script drops them from
consideration.}
    \item{\code{debugVRs}}{default is \code{NULL}.  If a vector of VR numbers is provided, the script will provide information about when the script drops them from
consideration.}
    \item{\code{debugMARFTripIDs}}{default is \code{NULL}.  If a vector of MARFIS trip IDs is provided, the script will provide information about when the script drops them from
consideration.}
    \item{\code{debugISDBTripIDs}}{default is \code{NULL}.  If a vector of ISDB trip IDs is provided, the script will provide information about when the script drops them from
consideration.  Trip "names" are typically in a format like "J18-1234" or "A18-1234A".}
    \item{\code{debugISDBTripNames}}{default is \code{NULL}.  If a vector of ISDB trip names is provided, the script will provide information about when the script drops them from
consideration.}
  }}
}
\value{
list of objects, including marfis data, isdb data, information for matching isdb
and marfis data, and a summary of bycatch
}
\description{
This function is a generic wrapper that facilitates extracting fleet specific information using the data provided by fleet-specific wrappers.
}
\examples{
\dontrun{
test <- fleet_(fleet="POLLOCK_MOBILE", area="WESTERN", gearSpecs="SMALL" )
 }
}
\seealso{
Other coreFuncs: 
\code{\link{get_fleet}()},
\code{\link{get_isdb}()},
\code{\link{get_marfis}()},
\code{\link{set_defaults}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
