% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_redfish.R
\name{fleet_redfish}
\alias{fleet_redfish}
\title{fleet_redfish}
\usage{
fleet_redfish(unit = NULL, useLocal = NULL, year = NULL, ...)
}
\arguments{
\item{unit}{default is \code{NULL}.  Valid selections include \code{"UNIT2"} and \code{"UNIT3"}}

\item{useLocal}{default is \code{NULL}.  Valid selections are  \code{"TRUE"} and \code{"FALSE"}
if \code{TRUE}, then the following parameters are necessary:
' \itemize{
  \item \code{data.dir} - This is a path on your computer to where your .RData files are stored.
  }
if \code{FALSE}, then the following parameters are necessary:
' \itemize{
  \item \code{oracle.username} - your existing oracle login name
  \item \code{oracle.password} - your existing oracle password
  \item \code{oracle.dsn} - usually "PTRAN" - the name of the database you're connecting to
  \item \code{usepkg} - default is \code{"rodbc"}, but \code{"roracle"} is also valid
  }}

\item{year}{default is \code{NULL}.  This can be a 4 digit year to extract that years' data. Alternatively,
\code{dateStart} and \code{dateEnd} can be entered as "YYYY-MM-DD" to extract longer or shorter windows of data}

\item{...}{
  Arguments passed on to \code{\link[=set_defaults]{set_defaults}}
  \describe{
    \item{\code{lics}}{default is \code{'all'}}
    \item{\code{gearSpecs}}{default is \code{'all'}}
    \item{\code{area}}{default is \code{'all'}}
    \item{\code{marfSpp}}{default is \code{'all'}}
    \item{\code{vessLen}}{default is \code{'all'}}
    \item{\code{useDate}}{default is \code{'LANDED_DATE'}}
    \item{\code{returnFleet}}{default is \code{TRUE}}
    \item{\code{returnMARFIS}}{default is \code{TRUE}}
    \item{\code{returnISDB}}{default is \code{TRUE}}
    \item{\code{returnBycatch}}{default is \code{TRUE}}
    \item{\code{returnLocations}}{default is \code{TRUE}}
    \item{\code{useReportedNAFO}}{default is \code{TRUE}}
    \item{\code{areas}}{default is \code{'NAFOSubunits_sf'}}
    \item{\code{areasField}}{default is \code{'NAFO_1'}}
    \item{\code{dateStart}}{default is \code{NULL}}
    \item{\code{dateEnd}}{default is \code{NULL}}
    \item{\code{keepSurveyTrips}}{default is \code{FALSE}}
    \item{\code{matchMarfis}}{default is \code{TRUE}}
    \item{\code{matchMaxDayDiff}}{default is \code{15}}
    \item{\code{dropUnmatchedISDB}}{default is \code{TRUE}}
    \item{\code{data.dir}}{default is \code{'file.path(getwd(), "data")'}}
    \item{\code{oracle.username}}{default is \code{'_none_'}}
    \item{\code{oracle.password}}{default is \code{'_none_'}}
    \item{\code{oracle.dsn}}{default is \code{'_none_'}}
    \item{\code{usepkg}}{default is \code{'rodbc'}}
    \item{\code{debugISDBTrips}}{default is \code{'_none_'}}
    \item{\code{HS}}{default is \code{FALSE}}
    \item{\code{debug}}{default is \code{FALSE}}
  }}
}
\value{
list of objects, including marfis data, isdb data, information for matching isdb
and marfis data, and a summary of bycatch, specifically:
#' a list item containing:
\itemize{
  \item \code{FLEET} - This is a dataframe of the unique combinations of (MARFIS) LICENCE_ID, VR_NUMBER and GEAR_CODE that
  was found for this fleet during the specified period
  \item \code{FLEET_ACTIVITY} - This is a dataframe of identifiers for all of the (MARFIS) fishing activity undertaken
  by vessels of this fleet during the specified period (i.e. LICENCE_ID, PRO_SPC_INFO_ID, LOG_EFRT_STD_INFO_ID, GEAR_CODE,
  MON_DOC_ID, VR_NUMBER, and several dates associated with the trip)
  \item \code{marf} - This is a list of 3 sets of information for the commercial catch data (i.e. marfis)-
  the trips, the sets, and a special dataframe containing information that can be used to link
  the commercial data to the ISDB data
  \item \code{isdb} - This is a list of 4 data objects - 2 of which are all of the discovered ISDB data
  TRIPS and SETS for the fleet, as well as the TRIPS and SETS from the observer data that were
  sucessfully matched with the MARFIS data
  \item \code{bycatch} - This is a dataframe with the various species that were observed during observed
  trips.  For each species, the estimated number caught, the estimated kept wt (kgs) and the
  estimated discarded wt(kg) are all captured
  }
}
\description{
This function is a wrapper function that facilitates extracting information for the redfish fleets.
}
\note{
Hardcoded parameters for this fleet are as follows:
\itemize{
  \item \code{marfSpp} = 120
}
}
\examples{
\dontrun{
db <- fleet_redfish(unit = "UNIT2",
                    useLocal = F,
                    oracle.username = "<name>",
                    oracle.password="<password>",
                    oracle.dsn="PTRAN",
                    usepkg = "roracle"
                    )
local <- fleet_redfish(unit = "UNIT2",
                       useLocal = T,
                       data.dir = "c:/data_folder"
                      )
                       }
}
\seealso{
Other fleets: 
\code{\link{fleet_}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
