% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleet_surfclam.R
\name{fleet_surfclam}
\alias{fleet_surfclam}
\title{fleet_surfclam}
\usage{
fleet_surfclam(useLocal = NULL, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=set_defaults]{set_defaults}}
  \describe{
    \item{\code{lics}}{default is an empty dataframe\code{'all'}}
    \item{\code{gearSpecs}}{default is \code{'all'}}
    \item{\code{area}}{default is \code{'all'}}
    \item{\code{marfSpp}}{default is \code{'all'}}
    \item{\code{vessLen}}{default is \code{'all'}}
    \item{\code{useDate}}{default is \code{'LANDED_DATE'}}
    \item{\code{returnFleet}}{default is \code{TRUE}}
    \item{\code{returnMARFIS}}{default is \code{TRUE}}
    \item{\code{returnISDB}}{default is \code{TRUE}}
    \item{\code{returnBycatch}}{default is \code{TRUE}}
    \item{\code{returnLocations}}{default is \code{TRUE}}
    \item{\code{useReportedNAFO}}{default is \code{TRUE}}
    \item{\code{manual_fleet}}{default is \code{FALSE}}
    \item{\code{areas}}{default is \code{'NAFOSubunits_sf'}}
    \item{\code{areasField}}{default is \code{'NAFO_1'}}
    \item{\code{dateStart}}{default is \code{NULL}}
    \item{\code{dateEnd}}{default is \code{NULL}}
    \item{\code{year}}{default is \code{NULL}}
    \item{\code{keepSurveyTrips}}{default is \code{FALSE}}
    \item{\code{matchMarfis}}{default is \code{TRUE}}
    \item{\code{matchMaxDayDiff}}{default is \code{15}}
    \item{\code{dropUnmatchedISDB}}{default is \code{TRUE}}
    \item{\code{data.dir}}{default is \code{'file.path(getwd(), "data")'}}
    \item{\code{oracle.username}}{default is \code{'_none_'}}
    \item{\code{oracle.password}}{default is \code{'_none_'}}
    \item{\code{oracle.dsn}}{default is \code{'_none_'}}
    \item{\code{usepkg}}{default is \code{'rodbc'}}
    \item{\code{useLocal}}{default is \code{FALSE}}
    \item{\code{quietly}}{default is \code{TRUE}}
    \item{\code{debugISDBTrips}}{default is \code{'_none_'}}
    \item{\code{HS}}{default is \code{FALSE}}
  }}

\item{fleetOnly}{default is \code{TRUE}.  If TRUE, this function will return information about this fleet, specifically,
a list item containing:
\itemize{
  \item \code{FLEET} - This is a dataframe of the unique combinations of (MARFIS) LICENCE_ID, VR_NUMBER and GEAR_CODE that
  was found for this fleet during the specified period
  \item \code{FLEET_ACTIVITY} - This is a dataframe of identifiers for all of the (MARFIS) fishing activity undertaken
  by vessels of this fleet during the specified period (i.e. LICENCE_ID, PRO_SPC_INFO_ID, LOG_EFRT_STD_INFO_ID, GEAR_CODE,
  MON_DOC_ID, VR_NUMBER, and several dates associated with the trip)
}
 If FALSE, this function will return the full suite of Mar.bycatch outputs.  So, in addition to the fleet information above,
 it will also return:
\itemize{
  \item \code{marf} - This is a list of 3 sets of information for the commercial catch data (i.e. marfis)-
  the trips, the sets, and a special dataframe containing information that can be used to link
  the commercial data to the ISDB data
  \item \code{isdb} - This is a list of 4 data objects - 2 of which are all of the discovered ISDB data
  TRIPS and SETS for the fleet, as well as the TRIPS and SETS from the observer data that were
  sucessfully matched with the MARFIS data
  \item \code{bycatch} - This is a dataframe with the various species that were observed during observed
  trips.  For each species, the estimated number caught, the estimated kept wt (kgs) and the
  estimated discarded wt(kg) are all captured
  }}
}
\value{
list of objects, including marfis data, isdb data, information for matching isdb
and marfis data, and a summary of bycatch
}
\description{
This function is a wrapper function that facilitates extracting information for the surfclam fleet.
}
\note{
Hardcoded parameters for this fleet are as follows:
\itemize{
  \item \code{marfSpp} = 608
}
}
\examples{
\dontrun{
Surfclam <- fleet_surfclam(data.dir = "C:/myData")
                       }
}
\seealso{
Other fleets: 
\code{\link{fleet_haddock}()},
\code{\link{fleet_halibut}()},
\code{\link{fleet_pollock}()},
\code{\link{fleet_redfish}()},
\code{\link{fleet_scallop}()},
\code{\link{fleet_silverhake}()},
\code{\link{fleet_snowcrab}()},
\code{\link{fleet_winterflounder}()},
\code{\link{fleet_}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
