% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bycatch.R
\name{get_bycatch}
\alias{get_bycatch}
\title{get_bycatch}
\usage{
get_bycatch(isTrips = NULL, marfSpID = NULL, ...)
}
\arguments{
\item{isTrips}{default is \code{NULL} This is a vector of trip_ids from ISDB}

\item{marfSpID}{default is \code{NULL}.  This is the Marfis species ID for the targeted species}

\item{...}{other arguments passed to methods}
}
\value{
returns a dataframe of the bycatch information for the requested species (using the requested)
marfis species code and supplie results from get_isdb
}
\description{
This function takes a vector of ISDB trip IDs and a marfis species code
and generates a dataframe of all of the bycatche species from the ISDB db
}
\note{
This function can accept many parameters.
}
\seealso{
Other simpleproducts: 
\code{\link{calc_coveragio}()},
\code{\link{get_vmstracks}()},
\code{\link{plot_bycatch}()},
\code{\link{plot_coverage_time}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
