% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fleet.R
\name{get_fleet}
\alias{get_fleet}
\title{get_fleet}
\usage{
get_fleet(...)
}
\arguments{
\item{...}{other arguments passed to methods}
}
\value{
returns a data.frame with the following columns:
\itemize{
  \item \code{MON_DOC_ID}
  \item \code{GEAR_CODE }
  \item \code{LICENCE_ID}
  \item \code{PRO_SPC_INFO_ID}
  \item \code{LOG_EFRT_STD_INFO_ID         }
  \item \code{T_DATE1}
  \item \code{T_DATE2}
  \item \code{MD_CODE}
  \item \code{VR_NUMBER}
  }
}
\description{
This function extracts all of the Vessel/Licence combinations
associated with a particular fleet for a particular date range.
}
\seealso{
Other coreFuncs: 
\code{\link{fleet_}()},
\code{\link{getED}()},
\code{\link{getHIC}()},
\code{\link{getHOC}()},
\code{\link{getPS}()},
\code{\link{get_isdb}()},
\code{\link{get_isdb_sets}()},
\code{\link{get_isdb_trips}()},
\code{\link{get_marfis}()},
\code{\link{get_marfis_sets}()},
\code{\link{set_defaults}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
