% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MARFIS_local.R
\name{get_MARFIS_local}
\alias{get_MARFIS_local}
\title{get_MARFIS_local}
\usage{
get_MARFIS_local(thisFleet = NULL, ...)
}
\arguments{
\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from
\code{Mar.bycatch::get_fleet()}}

\item{dateStart}{default is \code{NULL}. This is the start date (YYYY-MM-DD)
of the window of time you want to look at.}

\item{dateEnd}{default is \code{NULL}. This is the end date (YYYY-MM-DD)
of the window of time you want to look at.  If this is left blank, 1 year of
data will be returned.}

\item{useDate}{default is \code{"fished"}. Some MARFIS tables have 2 different dates
that are used for recording when fishing activity took place.  One is "DATE_FISHED",
and the other is "LANDED_DATE". If useDate = "fished", the DATE_FISHED field will be used for
subsetting data by date.  Any other value will result in the use of "LANDED_DATE" instead.}

\item{marfSpp}{default is \code{NULL}.  This is the marfis species code for the species you want
records for. There are literally hundreds of codes, but here are some of the more commonly used:
\itemize{
\item 100 = Cod
\item 110 = Haddock
\item 120 = Redfish
\item 130 = Halibut
\item 141 = Yellowtail Flounder
\item 142 = Witch Flounder
\item 143 = Winter Flounder
\item 144 = Turbot
\item 170 = Pollock
\item 172 = Silver Hake
\item 200 = Herring
\item 251 = Swordfish
\item 608 = Surf Clam
\item 619 = Sea Cucumber
\item 623 = Sea Scallop
\item 700 = Lobster
}}

\item{nafoCode}{default is \code{NULL}. This is a vector of NAFO AREAS (MARFIS) that will be
used to limit the fleet to.  If this is left as NULL, a popup will allow the user to select
valid values from a list. Codes can use '%' as a wildcard to ensure that all nafo areas that start
with the specified code are returned.  For example c('5%') would return all of the areas within
5Z and 5Y, c('5Z%'), would only return the areas within 5Z (e.g. 5ZE, 5ZU,...), and c('5Z') (no
wildcard) would only return 5Z - no subareas.}

\item{vessLen}{default is \code{NULL}.  This is a vector of vessel lengths.  If it is not NULL or
"all", it will be used to restrict vessels by their size.  If you wanted all vessels up to and
including 45 feet, you might enter a value of \code{seq(0,45,1)}.}
}
\value{
returns a list with 2 dataframes - "trips", and "sets".
"trips" conta ins all of the information necessary for identifying a trip
within MARFIS, as well as associated information about the trip
from the HAIL_OUTS and HAIL_IN_CALLS tables (e.g. confirmation numbers).
"sets" contains information about individual fishing activities, including
locations, dates, durations, gear amount, etc..
}
\description{
This function extracts all of the MARFIS records for vessels with
particular combinations of VR_NUMBER and LICENCE_ID  for a given date range.
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS_remote}()},
\code{\link{get_OBS_local}()},
\code{\link{get_OBS_remote}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet_local}()},
\code{\link{get_fleet_remote}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
