% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfis.R
\name{get_marfis_sets}
\alias{get_marfis_sets}
\title{get_marfis_sets}
\usage{
get_marfis_sets(log_efrt = NULL, ...)
}
\arguments{
\item{log_efrt}{default is \code{NULL}.}

\item{...}{other arguments passed to methods}

\item{extract_user}{default is \code{NULL}.  This parameter can be used with
\code{extract_computer} to load encypted data files extracted by another user
and/or computer}

\item{extract_computer}{default is \code{NULL}.  This parameter can be used with
\code{extract_user} to load encypted data files extracted by another user
and/or computer}
}
\value{
a dataframe
}
\description{
This function e
}
\seealso{
Other coreFuncs: 
\code{\link{fleet_}()},
\code{\link{getED}()},
\code{\link{getHIC}()},
\code{\link{getHOC}()},
\code{\link{getPS}()},
\code{\link{get_fleet}()},
\code{\link{get_isdb}()},
\code{\link{get_isdb_sets}()},
\code{\link{get_isdb_trips}()},
\code{\link{get_marfis}()},
\code{\link{set_defaults}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
