% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OBS.R
\name{get_OBS}
\alias{get_OBS}
\title{get_OBS}
\usage{
get_OBS(thisFleet = NULL, get_MARFIS = NULL, ...)
}
\arguments{
\item{thisFleet}{default is \code{NULL}. This is a dataframe that must include
the columns "LICENCE_ID" and "VR_NUMBER".  It can take the results from
\code{Mar.bycatch::get_fleet()}}

\item{get_MARFIS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_MARFIS()} function - it contains dataframes of both the
trip and set information from MARFIS related to the specified fleet}

\item{...}{other arguments passed to methods}
}
\value{
returns a list with 2 dataframes - "OBS_TRIPS", and "OBS_SETS".
"OBS_TRIPS" contains information for the observer trips, while "OBS_SETS" contains information
about the observer sets.
}
\description{
This function extracts all of the records from the observer database that are
within a particular date range, and if a fleet is provided, it will also limit the results to
those vessels with particular combinations of VR and licence.
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
