% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vmstracks.R
\name{get_vmstracks}
\alias{get_vmstracks}
\title{title get_vmstracks}
\usage{
get_vmstracks(data = NULL, ...)
}
\arguments{
\item{data}{default is \code{NULL}. This is the entire output from any of the fleet wrappers.}

\item{...}{other arguments passed to methods
You will need to provide the following parameter to enable a connection to oracle:
\itemize{
  \item \code{cxn} A valid Oracle connection object. This parameter allows you to
pass an existing connection, reducing the need to establish a new connection
within the function. If provided, it takes precedence over the connection-
related parameters.
}}
}
\value{
returns a dataframe of the VMS data.  The OBS field contains a value>0 if the trip was observed.
}
\description{
This function takes the results from get_marfis() and get_isdb()
and extracts the relevant VMS tracks, and flags whether or not each was observed
}
\examples{
\dontrun{
redfishresults <- fleet_redfish(unit=3, year = "2017", useLocal=T)
redfishVMS<-get_vmstracks(data = redfishresults,
                          cxn  = <valid Oracle Connection>)
                          }
}
\seealso{
Other simpleproducts: 
\code{\link{summarize_locations}()},
\code{\link{summarizer}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
