% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vmstracks.R
\name{get_vmstracks}
\alias{get_vmstracks}
\title{title get_vmstracks}
\usage{
get_vmstracks(get_marfis = NULL, get_isdb = NULL, ...)
}
\arguments{
\item{get_marfis}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_marfis()} function - it contains dataframes of both the
trip and set information from MARFIS}

\item{get_isdb}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_isdb()} function - it contains dataframes of both the
trip and set information from the ISDB database.}

\item{...}{other arguments passed to methods}
}
\value{
returns a dataframe of the VMS data.  The OBS field contains a value>0 if the trip was observed.
}
\description{
This function takes the results from get_marfis() and get_isdb()
and extracts the relevant VMS tracks, and flags whether or not each was observed
}
\seealso{
Other simpleproducts: 
\code{\link{calc_coveragio}()},
\code{\link{get_bycatch}()},
\code{\link{plot_bycatch}()},
\code{\link{plot_coverage_time}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
