% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_VMSTracks.R
\name{get_VMSTracks}
\alias{get_VMSTracks}
\title{title get_VMSTracks}
\usage{
get_VMSTracks(
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  usepkg = "roracle",
  get_MARFIS = NULL,
  get_OBS = NULL,
  quietly = FALSE
)
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file),
this can be left and that value will be used.  If a value for this is
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{get_MARFIS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_MARFIS()} function - it contains dataframes of both the
trip and set information from MARFIS}

\item{get_OBS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_OBS()} function - it contains dataframes of both the
trip and set information from the observer database.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}
}
\value{
returns a dataframe of the VMS data.  The OBS field contains a value>0 if the trip was observed.
}
\description{
This function takes the results from get_MARFIS() and get_OBS()
and extracts the relevant VMS tracks, and flags whether or not each was observed
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS_local}()},
\code{\link{get_MARFIS_remote}()},
\code{\link{get_OBS_local}()},
\code{\link{get_OBS_remote}()},
\code{\link{get_fleet_local}()},
\code{\link{get_fleet_remote}()},
\code{\link{match_sets}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
