% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vmstracks.R
\name{get_vmstracks}
\alias{get_vmstracks}
\title{title get_vmstracks}
\usage{
get_vmstracks(get_marfis = NULL, get_isdb = NULL, ...)
}
\arguments{
\item{get_marfis}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_marfis()} function - it contains dataframes of both the
trip and set information from MARFIS}

\item{get_isdb}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_isdb()} function - it contains dataframes of both the
trip and set information from the ISDB database.}

\item{...}{other arguments passed to methods
It is likely that you will need to provide values for the following 4
parameters to enable a connection to oracle:
\itemize{
  \item \code{oracle.username} This is your username for accessing oracle objects.
  \item \code{oracle.password} This is your password for accessing oracle objects.
  \item \code{oracle.dsn} This is your dsn/ODBC identifier for accessing oracle objects.
  \item \code{usepkg} This indicates whether the connection to Oracle should
  use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier
  to setup, but roracle will extract data more quickly.
}}
}
\value{
returns a dataframe of the VMS data.  The OBS field contains a value>0 if the trip was observed.
}
\description{
This function takes the results from get_marfis() and get_isdb()
and extracts the relevant VMS tracks, and flags whether or not each was observed
}
\seealso{
Other simpleproducts: 
\code{\link{quickMap}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
