% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marfisSummarizer.R
\name{marfisSummarizer}
\alias{marfisSummarizer}
\title{marfisSummarizer}
\usage{
marfisSummarizer(
  data = NULL,
  tonnes = F,
  byGr = TRUE,
  bySpp = TRUE,
  byYr = TRUE,
  byNAFO = FALSE
)
}
\arguments{
\item{data}{default is \code{NULL}. This is the entire output from any of the fleet wrappers.}

\item{byGr}{default is \code{TRUE}. If TRUE, the summed weights and numbers of unique trips and
licence_ids will will be broken down by all identified gear types. If FALSE,
weights will be summed, irrespective of gear.}

\item{bySpp}{default is \code{TRUE}. If TRUE, the summed weights and numbers of unique trips and
licence_ids will will be broken down by all identified landed species. If FALSE,
weights will be summed, irrespective of landed species.}

\item{byYr}{default is \code{TRUE}.   If TRUE, the summed weights and numbers of unique trips
and licence_ids will will be broken down by each year in which a landing was identified. If FALSE,
weights will be summed, irrespective of year. "Year" is taken from the end date of the associated
trip record.}

\item{byNAFO}{default is \code{FALSE}.   If TRUE, the summed weights and numbers of unique trips
and licence_ids will will be broken down by <reported> NAFO division in which a landing was
identified. If FALSE, weights will be summed, irrespective of <reported> NAFO division.}
}
\value{
a data frame
}
\description{
This is a simple tool that quickly breaks down the results of any of the fleet
wrappers to provide information on the number of trips, unique vessels, unique licences, and total
catch.  These results can be aggregated by YEAR, SPECIES_CODE, and GEAR TYPE.
}
\examples{
\dontrun{
summary <- marfisSummarizer(data = Halibut2017, byGr = TRUE)
       }
}
\seealso{
Other simpleproducts: 
\code{\link{get_vmstracks}()},
\code{\link{quickMap}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
