% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sets.R
\name{match_sets}
\alias{match_sets}
\title{match_sets}
\usage{
match_sets(
  get_obs = NULL,
  match_trips = NULL,
  marfMatch = NULL,
  marfSets = NULL,
  maxSetDiff_hr = 24,
  ...
)
}
\arguments{
\item{get_obs}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_obs()} function - it contains dataframes of both the
trip and set information from the observer database.}

\item{match_trips}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::match_trips()} function - it information related to how trips from
the two databases are matched.}

\item{marfMatch}{default is \code{NULL}. This is the MARF_MATCH output of the
\code{Mar.bycatch::get_marfis()} function - it contains dataframes of both the
trip and set information from MARFIS}

\item{marfSets}{default is \code{NULL}. This is the MARF_SETS output of the
\code{Mar.bycatch::get_marfis()} function - it contains information about the MARFIS sets.}

\item{maxSetDiff_hr}{default is \code{24}.  This is how many hours are allowed between
reported Observer and MARFIS sets before.  Sets differing by more than this time span
will never be matched.}

\item{...}{other arguments passed to methods}
}
\value{
a list containing a single dataframe - "MAP_OBS_MARFIS_SETS"
}
\description{
This function takes the results from get_marfis(), get_obs()
and match_trips(), and attempts to match the sets for each trip.
}
\seealso{
Other fleets: 
\code{\link{calc_coverage}()},
\code{\link{get_fleet}()},
\code{\link{get_marfis}()},
\code{\link{get_obs}()},
\code{\link{get_vmstracks}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
