% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sets.R
\name{match_sets}
\alias{match_sets}
\title{match_sets}
\usage{
match_sets(
  get_MARFIS = NULL,
  get_OBS = NULL,
  match_trips = NULL,
  maxSetDiff_hr = 24,
  quietly = F
)
}
\arguments{
\item{get_MARFIS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_MARFIS()} function - it contains dataframes of both the
trip and set information from MARFIS}

\item{get_OBS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_OBS()} function - it contains dataframes of both the
trip and set information from the observer database.}

\item{match_trips}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::match_trips()} function - it information related to how trips from
the two databases are matched.}

\item{maxSetDiff_hr}{default is \code{24}.  This is how many hours are allowed between
reported Observer and MARFIS sets before.  Sets differing by more than this time span
will never be matched.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}
}
\value{
a list containing a single dataframe - "MAP_OBS_MARFIS_SETS"
}
\description{
This function takes the results from get_MARFIS(), get_OBS()
and match_trips(), and attempts to match the sets for each trip.
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS}()},
\code{\link{get_OBS}()},
\code{\link{get_fleet}()},
\code{\link{match_trips}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
