% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_trips.R
\name{match_trips}
\alias{match_trips}
\title{match_trips}
\usage{
match_trips(get_MARFIS = NULL, get_OBS = NULL, quietly = FALSE)
}
\arguments{
\item{get_MARFIS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_MARFIS()} function - it contains dataframes of both the
trip and set information from MARFIS}

\item{get_OBS}{default is \code{NULL}. This is the list output by the
\code{Mar.bycatch::get_OBS()} function - it contains dataframes of both the
trip and set information from the observer database.}

\item{quietly}{default is \code{FALSE}.  This indicates whether or not
information about the matching process should be shown.}
}
\value{
returns a list with 3 dataframes - MAP_OBS_MARFIS_TRIPS, MATCH_ISSUES, UNMATCHABLE & MATCH_SUMMARY
\itemize{
\item "MAP_OBS_MARFIS_TRIPS" - contains the TRIP_IDs from MARFIS and OBSERVER, and
an additional feld that indicates all of the ways the match was attained.
\item "MATCH_ISSUES" - contains any trips that were matched, but had at least one invalid/
unmatchable entry due to something like a bad trip name.
\item "UNMATCHABLE" - contains any records from both Marfis and Observer that could not be
matched to the other database using any method.
}
}
\description{
This function takes the results from get_MARFIS() and get_OBS()
and attempts to match trips based on:
\itemize{
\item 1 = MARFIS confirmation numbers
\item 2 = Observer TRIP names (e.g. J18-0000)*
\item 3 = Correct combination of VRN and LICENCE and appropriate date range
}
* - only the alphanumeric characters of the trip names are used (e.g.
"J18-0000B" becomes "J180000B").
}
\seealso{
Other fleets: 
\code{\link{calc_Coverage}()},
\code{\link{get_MARFIS}()},
\code{\link{get_OBS}()},
\code{\link{get_VMSTracks}()},
\code{\link{get_fleet}()},
\code{\link{match_sets}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{fleets}
