% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMap.R
\name{quickMap}
\alias{quickMap}
\title{quickMap}
\usage{
quickMap(data = NULL, vms = NULL, nafo = TRUE)
}
\arguments{
\item{data}{default is \code{NULL}. This is the entire output from any of the fleet wrappers.}

\item{vms}{default is \code{NULL}. This is optional, but can be the output from \code{get_vmstracks()}.}

\item{nafo}{default is \code{TRUE}. This indicates whether or not the NAFO areas should be plotted.}
}
\value{
a leaflet map.
}
\description{
This function generates a simple leaflet plot for the output from Mar.bycatch
functions.
}
\examples{
\dontrun{
redfishresults <- fleet_redfish(unit=3, year = "2017", useLocal=T,data.dir="c:/data/")
redfishVMS<-get_vmstracks(get_marfis = redfishresults$marf, get_isdb = redfishresults$isdb,
         useLocal=F, oracle.username = "me", oracle.password = "mypassword",
         oracle.dsn="PTRAN", usepkg='roracle')
quickMap(data=redfishresults, vms= redfishVMS, nafo=TRUE)
                                }
}
\seealso{
Other simpleproducts: 
\code{\link{get_vmstracks}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{simpleproducts}
