% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{set_defaults}
\alias{set_defaults}
\title{set_defaults}
\usage{
set_defaults(
  lics = "all",
  gearSpecs = "all",
  area = "all",
  marfGear = "all",
  marfSpp = "all",
  isdbSpp = "all",
  tripcd_id = NULL,
  returnMARFIS = T,
  returnISDB = T,
  areaFile = "NAFOSubunits_sf",
  areaFileField = "NAFO_1",
  nafoDet = 2,
  dateStart = NULL,
  dateEnd = NULL,
  year = NULL,
  keepSurveyTrips = TRUE,
  maxTripDiff_Hr = 48,
  maxSetDiff_Hr = 48,
  dropUnmatchedISDB = TRUE,
  manualMatch = FALSE,
  data.dir = file.path(getwd(), "data"),
  oracle.username = "_none_",
  oracle.password = "_none_",
  oracle.dsn = "PTRAN",
  usepkg = "roracle",
  useLocal = FALSE,
  quietly = TRUE,
  debugLics = NULL,
  debugVRs = NULL,
  debugMARFTripIDs = NULL,
  debugISDBTripIDs = NULL,
  debugISDBTripNames = NULL,
  debug = FALSE,
  ...
)
}
\arguments{
\item{lics}{default is \code{'all'}.  Fleet wrappers will populate it with a dataframe specifying
the licence types, subtypes, gear and licence species valid
for a particular fleet.  This information comes from  \code{Mar.fleets::LIC_CORE}.}

\item{gearSpecs}{default is \code{'all'}. Fleet wrappers may populate it with a dataframe
specifying the particular specifications for a fleets' gear.
For example, certain fleets must use particular mesh sizes or shapes. This information comes from
\code{Mar.fleets::LIC_GEAR_SPECS}.}

\item{area}{default is \code{'all'}.  Certain fleets are only licenced to fish in certain areas.
This information comes from \code{Mar.fleets::licAreas}.}

\item{marfGear}{default is \code{'all'}, but all wrappers have (overwritable) fleet-specific
values.  This is a vector of MARFIS gear codes known to have caught this species.}

\item{marfSpp}{default is \code{'all'}. The marfis species code, usually sent by the fleet
wrapper.}

\item{isdbSpp}{default is \code{'all'}. The ISDB species code, usually sent by the fleet wrapper}

\item{tripcd_id}{default is \code{NULL}.  If a tripcd_id from ISDB is provided, all matting
records will be examined for matches}

\item{returnMARFIS}{default is \code{TRUE}. Do you want a list object containing marfis trip and
set information as part of your results?}

\item{returnISDB}{default is \code{TRUE}. Do you want a list object containing isdb trip and set
information as part of your results? (requires \code{returnMARFIS = T})}

\item{areaFile}{default is \code{'NAFOSubunits_sf'}.  This is used to identify which areas to
check the trips and sets against. By default,
Mar.data::NAFOSubunits_sf is ued, but any objects in Mar.data could be used.}

\item{areaFileField}{default is \code{'NAFO_1'}. This is a field within the \code{areas} object
which specifies exactly which field of the areas object data
should be compared against.}

\item{nafoDet}{default is \code{2}, but values between \code{1} and \code{4} are acceptable. This
specifies the level of detail that will be used in the summarized locations table.  Using the
default value of 2, trips and sets will be summarized by areas such as "4X", "4V" and "5Z" (i.e 2
characters).  If set to "1", areas would be more general  (e.g. "3", "4", "5"; i.e. 1 character),
while a value like 4 would summarize the trips and sets into very specific NAFO subunits (e.g.
"3PSA","4VSB" and "5ZEM")}

\item{dateStart}{default is \code{NULL}.  This identifies the start date of the search window, in
the format \code{YYYY-MM-DD} (e.g. \code{dateStart = "2019-02-19"}).
If no \code{dateEnd} is provided, the window will be for 1 year (i.e, 365 days).  ,}

\item{dateEnd}{default is \code{NULL} format is \code{YYYY-MM-DD}, e.g. \code{dateEnd = "2019-02-19"}.
\code{dateEnd} must be associated with a valid entry of
\code{dateStart} to identify a block of time for a data extraction  (e.g. Jun18, 2018 -->
August 27, 2019).}

\item{year}{default is \code{NULL}.  \code{year} can be used if data should be extracted for an
entire calendar year (i.e. Jan 1 --> Dec 31).  The format is
\code{YYYY}, e.g. \code{year = 2018}.  \code{dateStart} takes priority over \code{year} if both
are provided.}

\item{keepSurveyTrips}{default is \code{TRUE}. Within the ISDB database are non-commercial,
survey trips.  Setting this to \code{FALSE} ensures these trips are dropped.}

\item{maxTripDiff_Hr}{default is \code{48}. Any MARFIS and ISDB trips that vary by more than the
# of days specified here will NOT be considered matches (on the basis of common Vessel, licence
and date).  They may still match on confirmation codes and/or trip names.}

\item{maxSetDiff_Hr}{default is \code{48}. Any MARFIS and ISDB sets that vary by more than the
# of hours specified here will NOT be considered matches.}

\item{dropUnmatchedISDB}{default is \code{TRUE}.}

\item{manualMatch}{default is \code{FALSE}. This parameter is only used when calling functions
from \code{manual_matcher()}.  It ensures that the functions work properly with its reduced
input format.}

\item{data.dir}{default is \code{'file.path(getwd(), "data")'}. Necessary for useLocal == T.
This is the path to a folder where your *.rdata files are stored.}

\item{oracle.username}{default is \code{'_none_'}. This is your username for accessing oracle
objects.}

\item{oracle.password}{default is \code{'_none_'}. This is your password for accessing oracle
objects.}

\item{oracle.dsn}{default is \code{'_none_'}.  This is your dsn/ODBC identifier for accessing
oracle objects.  Normally, the value should be "PTRAN"}

\item{usepkg}{default is \code{'roracle'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc can
be slightly easier to setup, but roracle will extract data faster.}

\item{useLocal}{default is \code{FALSE}. This specifies whether to run the script against local
data or against Oracle (requires network or VPN).
Depending on your value for \code{useLocal}, different values become necessary.
\itemize{
 \item{useLocal=TRUE} This implies that you have local data you want to use.
    \itemize{
      \item param \code{data.dir} required.  This is the path to your local data
    }
 \item{useLocal=FALSE} This implies that you have will query Oracle for the necessary data.
 Include all of the following:
  \itemize{
      \item param \code{oracle.username} required
      \item param \code{oracle.password} required
      \item param \code{oracle.dsn} required
      \item param \code{usepkg} required
    }
}}

\item{quietly}{default is \code{TRUE}. This specifies whether or not status messages should be
output to the console while the scripts run.}

\item{debugLics}{default is \code{NULL}.  If a vector of LICENCE_IDs is provided, the script will
provide information about when the script drops them from
consideration.}

\item{debugVRs}{default is \code{NULL}.  If a vector of VR numbers is provided, the script will
provide information about when the script drops them from
consideration.}

\item{debugMARFTripIDs}{default is \code{NULL}.  If a vector of MARFIS trip IDs is provided, the
script will provide information about when the script drops them from
consideration.}

\item{debugISDBTripIDs}{default is \code{NULL}.  If a vector of ISDB trip IDs is provided, the
script will provide information about when the script drops them from
consideration.  Trip "names" are typically in a format like "J18-1234" or "A18-1234A".}

\item{debugISDBTripNames}{default is \code{NULL}.  If a vector of ISDB trip names is provided,
the script will provide information about when the script drops them from
consideration.}

\item{debug}{default is \code{FALSE}. If TRUE, this parameter causes the package to run in
debug mode, providing much extraneous information.}

\item{...}{other arguments passed to methods}
}
\description{
This function provides all of the possible parameters understood by this package.
Some can be overwritten by \code{fleet_*} wrappers.
}
\details{
The desired date range can be specified multiple ways - either using \code{year} (alone),
\code{dateStart} (alone) or \code{dateStart} with \code{dateEnd} (explained below).
}
\seealso{
Other coreFuncs: 
\code{\link{fleet_}()},
\code{\link{getED}()},
\code{\link{getHIC}()},
\code{\link{getHOC}()},
\code{\link{getPS}()},
\code{\link{get_fleet}()},
\code{\link{get_isdb_sets}()},
\code{\link{get_isdb_trips}()},
\code{\link{get_isdb}()},
\code{\link{get_marfis_sets}()},
\code{\link{get_marfis}()},
\code{\link{summarize_locations}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
