% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_locations.R
\name{summarize_locations}
\alias{summarize_locations}
\title{summarize_locations}
\usage{
summarize_locations(get_isdb = NULL, get_marfis = NULL, ...)
}
\arguments{
\item{get_isdb}{default is \code{NULL}. This is the list output by the
\code{Mar.fleets::get_isdb()} function - it contains dataframes of both the
trip and set information from ISDB related to the specified fleet}

\item{get_marfis}{default is \code{NULL}. This is the list output by the
\code{Mar.fleets::get_marfis()} function - it contains dataframes of both the
trip and set information from MARFIS related to the specified fleet}

\item{...}{other arguments passed to methods}
}
\value{
returns a list with 1 or 2 dataframes. These dataframes indicate how much of
the fishing activity took place in various areas.  Values are categorized as both
"*_reported" and "*_calculated".  "Reported" values are based on the NAFO information
stored in MARFIS and ISDB.  These are available for MARFIS trips, MARFIS sets, and ISDB sets.
"Calculated" values are based on the coordinates associated with the data, and are only
available for the MARFIS and ISDB sets (not trips).
\itemize{
\item \code{NAFO} =  This is always provided.
\item \code{<Mar.data::object_name>} = This is only provided when a specific "areaFile" is
provided as a paramater.
}
}
\description{
This function takes the results from get_marfis() and get_isdb()
and produces a table showing the how the proportion of observed data varies
across areas.  Results for NAFO areas are always provided, as these areas are
stored in both MARFIS and ISDB databases.  If a custom area is specified (via the
\code{areaFile} parameter), the sets from both the MARFIS and ISDB data are analyzed
for that area too (as a separate dataframe).
}
\note{
\itemize{
\item \code{_SETS} Coordinates for MARFIS_* data comes from the db table "LOG_EFRT_STD_INFO"
(preferring *_ENT coordinates over *_DET ), and coordinates for ISDB_* data come from
"ISSETPROFILE" - using the first non-null pair found while checking from P1-P4 in order.
\item \code{Weird Coordinates} In addition to falling within a known area, data can also be
categorized any of the following:
\itemize{
\item \code{<missing coord>}  These are missing values for with LAT or LONG (or both)
\item \code{<impossible coord>}  These have latitudes or longitudes outside of the acceptable
ranges of 90 - -90  and 180 - -180, respectively
\item \code{<on boundary line>}  These coordinates fall exactly on a boundary line dividing 2
areas
\item \code{<LAND>} For NAFO areas, points that fall on land (according to
Mar.data::NAFOSubunitsLnd_sf) are indicated as such
\item \code{<outside known areas>} These are points that fall outside any of the known polygons
for the specified area. For non_NAFO areas, this can include land.
}
}
}
\seealso{
Other coreFuncs: 
\code{\link{fleet_}()},
\code{\link{get_fleet}()},
\code{\link{get_isdb}()},
\code{\link{get_marfis}()},
\code{\link{set_defaults}()}

Other simpleproducts: 
\code{\link{get_vmstracks}()},
\code{\link{quickMap}()},
\code{\link{summarizer}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{coreFuncs}
\concept{simpleproducts}
